### 2024-01-10 #6205 @editorialbot
[PRE REVIEW]: pySLM2: A full-stack python package for holographic beam shaping
**Submitting author:** @ldes89150 (<a href="http://orcid.org/0000-0002-7561-6833">Chung-You Shih</a>)
**Repository:** https://github.com/QITI/pySLM2
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.1.0
