### 2024-01-10 #6205 @kyleniemeyer
@ldes89150 it looks like the JOSS paper is located on the `paper.md` branch, is that correct?
