### 2024-01-10 #6205 @ldes89150
Hi @kyleniemeyer ,

Yes, it is in the `paper.md` branch. Do you need me to make any changes?
