### 2024-01-10 #6205 @kyleniemeyer
@ldes89150 we just needed to specify the branch, since our editorialbot looks for the paper in the main branch by default.

I'm not quite sure what is causing the error in building the PDF, though.
