### 2024-01-10 #6205 @kyleniemeyer
@openjournals/dev @arfon do you understand this error in the PDF build job?

```
Run bundle exec ruby /home/runner/work/_actions/xuanxu/paper-action/main/get_paper.rb
/home/runner/work/_actions/xuanxu/paper-action/main/vendor/bundle/ruby/3.2.0/gems/commonmarker-0.23.10/lib/commonmarker.rb:38:in `render_doc': text must be a String; got a NilClass! (TypeError)

      raise TypeError, "text must be a String; got a #{text.class}!" unless text.is_a?(String)
            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
	from /home/runner/work/_actions/xuanxu/paper-action/main/vendor/bundle/ruby/3.2.0/gems/theoj-1.9.0/lib/theoj/submission.rb:189:in `plaintext'
	from /home/runner/work/_actions/xuanxu/paper-action/main/vendor/bundle/ruby/3.2.0/gems/theoj-1.9.0/lib/theoj/submission.rb:100:in `all_metadata'
	from /home/runner/work/_actions/xuanxu/paper-action/main/vendor/bundle/ruby/3.2.0/gems/theoj-1.9.0/lib/theoj/submission.rb:78:in `metadata_info'
	from /home/runner/work/_actions/xuanxu/paper-action/main/vendor/bundle/ruby/3.2.0/gems/theoj-1.9.0/lib/theoj/submission.rb:57:in `article_metadata'
	from /home/runner/work/_actions/xuanxu/paper-action/main/get_paper.rb:25:in `<main>'
Error: Process completed with exit code 1.
```
