### 2024-01-11 #6205 @xuanxu
Probably the error comes from the branch being named `paper.md`. Git creates internal files with the names of every branch so there is a `paper.md` file that can confuse the bot if it finds it before the real paper file.
Renaming the branch should fix the error.
