### 2024-01-30 #6205 @HaoZeke
> @HaoZeke I can do a review from a general perspective on code quality, the paper, and so-on. But I do not have access to the hardware devices I believe the package is meant to control, and could not verify the functionality in a strict way

That makes sense to me, and thank you for clarifying. It seems like there is a software only component which would be better reviewed and the hardware control aspect would generally be very hard to verify.
