### 2024-01-30 #6205 @HaoZeke
> > hi @maurov 👋 would you be interested in and available to review this JOSS submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)
> > If not, could you recommend any potential reviewers?
>
> @HaoZeke thank you for the invitation. Unfortunately this contribution is out of my field of expertise and I do not know any addition potential reviewers for it.

Thanks for the prompt response. Hope to see you around on another review later :)
