### 2024-01-31 #6205 @HaoZeke
> Hi @HaoZeke , I can review it. However, with the same remark as the colleague: I do not have access to the hardware part to test this functionality.

Excellent, thank you for the prompt response, testing only the software component will be fine.
