### 2024-01-31 #6205 @HaoZeke
> > @HaoZeke I can do a review from a general perspective on code quality, the paper, and so-on. But I do not have access to the hardware devices I believe the package is meant to control, and could not verify the functionality in a strict way
>
> Hey @HaoZeke @brandondube ,
>
> Just want to clarify on the hardware part: Most part of the software and majority of the functionalities don't require the hardware. All the hardware related code is in `pySLM2/utils`.
>
> And @brandondube Thank you for your interest in reviewing the paper/repo!

@ldes89150, since we haven't found any reviewers for testing the hardware integration, would you be willing to add `[not-reviewed]` to the section describing the integration? If so, we can immediately go ahead and start the review.

Any suggestions for other reviewers for the hardware section would also be very welcome.
