### 2024-02-01 #6205 @HaoZeke
Sorry @sidihamady, I need to wait for @ldes89150 to confirm the `[not-reviewed]` label for the hardware section before I can start this one.
