### 2024-02-02 #6205 @ldes89150
@HaoZeke What exactly do I need to do?
I confirm not review for the hardware section (`pySLM2/utils`).
