### 2024-02-02 #6205 @ldes89150
I added it to the paper. Let me know if there is still something I need to do.
https://github.com/QITI/pySLM2/blob/paper/paper.md#hardware-controls-not-reviewed

Thanks!
