### 2024-02-02 #6315 @HaoZeke
@ldes89150, sorry about this, but after deliberations with @kyleniemeyer I believe it would be best to remove the section on Hardware Integration instead of leaving it in with `[not-reviewed]`. The rest of the review can proceed, just without that section.
