### 2024-02-02 #6315 @kyleniemeyer
@ldes89150 yes - we can include that section in the paper, if you are able to help us find an editor with access to the hardware.
