### 2024-02-08 #6315 @HaoZeke
Thanks a ton for taking time to review this!

>  Not sure how this works though, I read JOSS review guidelines, but where am I supposed to report my comments? Here?

The review checklist is generated here, and the final discussion is here (along with responses to the review) but we also recommend directly opening issues in the repository (mentioning that they're part of a JOSS review.
