### 2024-02-08 #6315 @wavefrontshaping
Hi,

I installed the package in a new environment.
I tried to build the docs using `make html`.

First, the dependencies for the docs are not installed when I installed the package, did I do something wrong?
I did not find instructions for making the docs in the repo (maybe I missed something), so I add the packages one by one when the errors appeared.
It may be useful to a requirement file in the docs folder.

Then, after installing the necessary packages, I run into an error:
```bash
Theme error:
no theme named 'furo' found (missing theme.conf?)
```
How do I solve that?
