### 2024-02-08 QITI/pySLM2#10 @ldes89150
Add docs building dependencies
Hey @wavefrontshaping,

Thank you for your suggestions in adding build dependencies for the docs in a file!

For now, it is listed in the github action.
https://github.com/QITI/pySLM2/blob/61cbc71516c5df131d9e5ce3ac5350dea177e2fd/.github/workflows/sphinx.yml#L30

You can also find the documentation built by the CI here.
