### 2024-02-08 #6315 @ldes89150
Hey @wavefrontshaping ,

Thank you so much for helping the review. I opened an issue on github for the problem you encountered.
https://github.com/QITI/pySLM2/issues/10
