### 2024-02-09 6d0be7f
Updated README and added requirements.txt for documentation
