+++ requirements.txt
@@ -1,0 +1,6 @@
+matplotlib
+furo
+m2r2
+sphinx
+sphinx-math-dollar
+sphinx_pyproject
