### 2024-02-09 432b290
modify github action. add pip install -r
