### 2024-02-09 c8ff963
update instruction to install doc builiding pkg in requirements.txt which is in docs folder instead of root folder
