### 2024-02-09 aea9b04
modified github workflow to install doc pkg from docs/requirements.txt
