### 2024-02-09 1e0ee1f
Updated README and added requirements.txt for documentation (#11)
* Updated README and added requirements.txt for documentation

* modify github action. add pip install -r

* Update requirements.txt

* update instruction to install doc builiding pkg in requirements.txt which is in docs folder instead of root folder

* modified github workflow to install doc pkg from docs/requirements.txt
