### 2024-02-09 QITI/pySLM2#10 @ldes89150
Addressed in https://github.com/QITI/pySLM2/pull/11
