### 2024-02-14 #6315 @sidihamady
Installation: The package is easy to install and the examples provided allow you to test it without problem. The only example that I can't test is obviously the one that uses the Luxbeam controller since I don't have access to the hardware.
