### 2024-02-16 #6315 @sidihamady
* the use of TensorFlow is a good choice. It can permit faster calculations and other possibilities. Unfortunately I test it only on CPU, without using GPU, then I have no quantitative comparison between calculations speed using GPU and CPU.
* the code is well written, documented and modular. Then it is easy to understand and to adapt to a specific application.
* The manuscript is well written, clear and the examples are presented coherently.

* In the "statement of need" section, authors highlight the advatange of holographic beam shaping using SLM, and then the need of their package. But are there other packages (in Python or not) with the same functionality? And if so, what else does pySML2 bring?
* The relative contribution of the three authors is not mentioned. It is assumed that the first author is the main one, but it is necessary to clarify the contributions.
