### 2024-02-17 QITI/pySLM2#13 @brandondube
Installation / GPU acceleration
This issue is part of my JOSS review at https://github.com/openjournals/joss-reviews/issues/6315

I created an environment to test your package with on Windows 11 and a system with an Nvidia 4080 GPU:

```sh
$ conda create -n pyslm2rev numpy scipy matplotlib tensorflow pytest
$ conda activate pyslm2rev
$ cd pySLM2
$ pip install -e .
```

Your tests passed:

<details><summary><code>pytest .</code> Output</summary><br><pre>
================================================= test session starts =================================================
platform win32 -- Python 3.10.13, pytest-7.4.0, pluggy-1.0.0
rootdir: C:\Users\brand\src\pySLM2
collected 41 items

test\test_analysis.py ..                                                                                         [  4%]
test\test_backend.py s                                                                                           [  7%]
test\test_profile.py ..................                                                                          [ 51%]
test\test_simulation.py ...s.                                                                                    [ 63%]
test\test_slm.py ..ss.....                                                                                       [ 85%]
test\test_util.py ssssss                                                                                         [100%]

================================================== warnings summary ===================================================
..\..\miniconda3\envs\pyslm2rev\lib\site-packages\tensorflow\python\framework\dtypes.py:246
  C:\Users\brand\miniconda3\envs\pyslm2rev\lib\site-packages\tensorflow\python\framework\dtypes.py:246: DeprecationWarning: `np.bool8` is a deprecated alias for `np.bool_`.  (Deprecated NumPy 1.24)
    np.bool8: (False, True),

..\..\miniconda3\envs\pyslm2rev\lib\site-packages\flatbuffers\compat.py:19
  C:\Users\brand\miniconda3\envs\pyslm2rev\lib\site-packages\flatbuffers\compat.py:19: DeprecationWarning: the imp module is deprecated in favour of importlib and slated for removal in Python 3.12; see the module's documentation for alternative uses
    import imp

..\..\miniconda3\envs\pyslm2rev\lib\site-packages\tensorboard\compat\tensorflow_stub\dtypes.py:326
  C:\Users\brand\miniconda3\envs\pyslm2rev\lib\site-packages\tensorboard\compat\tensorflow_stub\dtypes.py:326: DeprecationWarning: `np.bool8` is a deprecated alias for `np.bool_`.  (Deprecated NumPy 1.24)
    np.bool8: (False, True),

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
===================================== 31 passed, 10 skipped, 3 warnings in 18.83s =====================================
</pre></details>

Tensorflow did not recognize my GPU, even though CuPy does.  The version of tf that conda picked up was `tensorflow-2.10.0-mkl_py310hd99672f_0`.  I uninstalled tf with conda and installed it through pip, same issue:

```python
import tensorflow as tf
print("Num GPUs Available: ", len(tf.config.list_physical_devices('GPU')))
>>> Num GPUs Available:  0
```

The version of tf that pip picked up was tensorflow-intel-2.15.0.  To show that it is not an issue with my system, I installed CuPy into the same environment, and it worked just fine:

```
$ conda install -c conda-forge cupy
$ python
Python 3.10.13 | packaged by Anaconda, Inc. | (main, Sep 11 2023, 13:24:38) [MSC v.1916 64 bit (AMD64)] on win32
>>> import cupy as cp
>>> cp.ones(2)
array([1., 1.])
```

As a hail mary, I tried more-or-less the same in WSL2:
```
conda create -n pyslm2rev -c conda-forge numpy scipy matplotlib tensorflow pytest cupy
```

This grabs tf 2.12-mkl on python 3.9.  Same problem.

<details><summary>WSL2 tf failure to see GPU, CuPy working:</summary><br><pre>
$ python
Python 3.9.18 (main, Sep 11 2023, 13:21:18)
[GCC 11.2.0] :: Anaconda, Inc. on linux
>>> import tensorflow as tf
print("Num GPUs Available: ", len(tf.config.list_physical_devices('GPU')))2024-02-17 10:27:46.993035: I tensorflow/core/platform/cpu_feature_guard.cc:182] This TensorFlow binary is optimized to use available CPU instructions in performance-critical operations.
To enable the following instructions: SSE4.1 SSE4.2 AVX AVX2 AVX512F AVX512_VNNI AVX512_BF16 FMA, in other operations, rebuild TensorFlow with the appropriate compiler flags.
>>> print("Num GPUs Available: ", len(tf.config.list_physical_devices('GPU')))
Num GPUs Available:  0
>>> import cupy as cp
>>> cp.ones(2)
array([1., 1.])
</pre>

----

In your paper, you wrote:

> Under the hood, the package uses TensorFlow for numerical computations. By leveraging
TensorFlow, the package harnesses the power of GPUs for faster computation without the
need for code modification. This results in a significant speed-up for algorithms that are
computationally expensive but benefit from parallelization, such as many hologram generation
algorithms relying on iterative Fourier transformations.

While I understand that in theory you can use the same lines of code to work with either a CPU or GPU using tensorflow or other libraries of its ilk (pytorch, say), this has nor borne out in my case here, putting in considerable effort.  I do not have a personal bare metal linux box with a GPU to test with.

Have you demonstrated the claim in the paper?  If so, could you share the result?
