### 2024-02-17 QITI/pySLM2#14 @brandondube
Documentation
This issue is part of my JOSS review at https://github.com/openjournals/joss-reviews/issues/6315

The documentation review criteria that I can't check off right now are:

This criteria is missing both in your README, and your docs website

This criteria is present only in your README; please add it to the docs website

- [x] Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Please add a CONTRIBUTING.[md/txt] or likewise

----

One of the other criteria is to have at least API-level documentation, as well as examples.  You have both of these.  A nit-pick is that the API documentation is a single page (https://pyslm2.pages.dev/pySLM2) for the entire package, less the hardware module.  A reader may be better served by breaking this page into multiple; perhaps one for what you call profiles, one for simulation, and so on.

----

A second nit is that your documentation is fairly bare-bones.  For example, if I want to understand how to use your package to e.g. fit polynomials to data, I can find in the API documentation that this is a feature, but there is nothing that teaches me how to do so.  I may recommend [the documentation system](https://documentation.divio.com/) to create very user friendly docs
