### 2024-02-17 QITI/pySLM2#16 @brandondube
JOSS paper
This issue is part of my JOSS review at https://github.com/openjournals/joss-reviews/issues/6315

The review criteria for the paper are

- [ ] Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

The current draft of the paper can be outlined as follows

- holographic beam shaping with SLMs
- summary of pySLM2 (one sentence)
- technical detail on hologram algorithms (related, #15)
- tensorflow / gpus (related, #13)
- universal* interface to two SLM/DMD vendors' devices

I struggle with the `...diverse, non-specialist audience` portion of the criteria.  I also feel that the summary does not really "summarize" the software.  The one sentence is

> pySLM2 is a python package designed for full-stack control of SLMs for holographic beam
shaping, encompassing hologram generation, simulation, and hardware controls.

full-stack is not defined, one can maybe infer that it has something to do with both hardware control and (?) portion of software from the closing paragraph about universal control.  Spending this morning rummaging in your code, perhaps "summary" could/should touch on:

- the software can simulate the far-field / fourier plane of a DMD or SLM device
- the software can compute a command to apply to a DMD/SLM to achieve a given far-field complex E-field, given the field incident on the DMD/SLM
- likewise, the software can compute the command to generate a specific aberration correction
- the software includes numerous common decompositions in optics (HG, LG, SG, Zernike, ...)
- the software includes an abstraction of DMD/SLM vendor APIs, providing a universal interface to apply computed commands to hardware

- [ ] A statement of need: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?

Missing `...and its relation to other work?`

- [ ] State of the field: Do the authors describe how this software compares to other commonly-used packages?

no

- [x] Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

- [x] References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?
