### 2024-02-18 QITI/pySLM2#13 @ldes89150
Hey @brandondube

To install the GPU accelerated version, one can do `pip install tensorflow[and-cuda]`.
Though we link to the official TensorFlow's page for installation, I will add more documentation on the tensorflow's installation to make it more clear. Thank you very much for the feedback!

Colab provides environment with necessary dependencies installed. (make sure to choose the GPU runtime).
Here is an example, I run the tests with the GPU accelerated tensorflow.
https://colab.research.google.com/gist/ldes89150/db64b42aaa0edbef9eb583fd3e36a935/pyslm2.ipynb
