### 2024-02-19 5f8786e
move documentation building instructions in docs/README.md. Adding dependencies section to README.md in root folder
