+++ README.md
@@ -10,6 +10,14 @@ pySLM2
 
 * API Docs: https://pyslm2.pages.dev/
 
+Instructions to build documentation locally can be found in [`doc/README.md`](doc/README.md).
+
+
+Dependencies
+------------
+
+
+
 Installation
 ------------
 The dependencies of `pySLM2` in includes: `numpy`, `scipy`, `matplotlib`, and `tensorflow`.
@@ -28,15 +36,7 @@ Or, if you wish to edit the pySLM2 source code without re-installing each time
 ```
 pip install -e .
 ```
-To view pySLM2 documentation, some additional packages are required. In the pySLM2 directory, do
-```
-pip install -r docs/requirements.txt
-```
-This will install the necessary packages for building the documentation. Then  `cd` to the docs directory and in command line do 
-```
-make html
-```
-HTML files for the documentation will be generated and can be viewed in browser.
+
 
 Optional Dependencies for Hardware Controls
 -------------------------------------------
