### 2024-02-19 QITI/pySLM2#13 @brandondube
Google's provided installation documentation is here:

https://www.tensorflow.org/install

Which dictates to do as I did,

```
# Current stable release for CPU and GPU
pip install tensorflow
```

----

I added `tf.debugging.set_log_device_placement(True)` to the collab notebook and ran `examples/create_multiple_gaussian_beam.py` there.  It took ~16 sec, and tf indicated everything ran on the GPU.  It takes about two seconds on my machine, with CPU.  Understanding that the collab environment is quite different to my machine which is bare metal, I disabled the GPU in collab with `tf.config.set_visible_devices([], 'GPU')`.  The debug printing I added showed everything was being done on CPU, and the sample example took 7 seconds.  These timings are repeatable.

I am not sure how to confirm that by using tensorflow the calculations are faster on a GPU.

In general, you are satisfying the important criteria; your arrays are 32-bit floats, and large (1920 x 1920).  At these array sizes you should see a big speedup.  I ran `!nvidia-smi` and `!lscpu` to probe the collab environment, and it has an Nvidia T4 GPU (~8 Tflops in fp32).  The CPU comes up as model 6 family 79, which is Broadwell.  From the core clock and cache, it has to be an [E5-2699 v4](https://ark.intel.com/content/www/us/en/ark/products/91317/intel-xeon-processor-e5-2699-v4-55m-cache-2-20-ghz.html).  [Microway](https://www.microway.com/knowledge-center-articles/detailed-specifications-of-the-intel-xeon-e5-2600v4-broadwell-ep-processors/) lists this CPU as having lets say about 0.35 Tflops of compute, for all cores.  Since the VM has only one core / two threads, it has 1/22nd of that, about 0.016 Tflops.  The theoretical performance of the GPU is about 500x higher.

I installed CuPy into the same collab notebook (or tried to, it was already installed); `!pip install cupy-cuda12x`.  I then ran a test:

```python
import numpy as np
import cupy as cp

# old style numpy random api does not have a dtype= kwarg
acpu = np.random.rand(1024,1024).astype(np.float32)
bcpu = np.random.rand(1024,1024).astype(np.float32)

agpu = cp.random.rand(1024,1024).astype(cp.float32)
bgpu = cp.random.rand(1024,1024).astype(cp.float32)

%timeit acpu * np.exp(1j*bcpu)
37.1 ms ± 478 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

%timeit agpu * cp.exp(1j*bgpu)
72.6 µs ± 38.5 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

The ratio of these times is 511x, almost exactly in-line with the difference in their theoretical floating point performance.  So collab's hardware is very much in-line with my expectations.

----

As-is, I could not check the box for the review:

I can give you some pointers for how to optimize code for both CPU and GPU, as well as a small amount of code (~12 lines) that lets you hot swap your computational backend so you are not tied to numpy or tensorflow or cupy or pytorch or any other; provided they all use the same API, or same-enough; which they all do.
