### 2024-02-19 3d46d6b
more detailes on dependencies and tensorflow installation
