### 2024-02-21 QITI/pySLM2#18 @sidihamady
"Statement of need" section and contributions
[https://github.com/openjournals/joss-reviews/issues/6315#issuecomment-1948037721](https://github.com/openjournals/joss-reviews/issues/6315#issuecomment-1948037721)

In the "statement of need" section, authors highlight the advantage of holographic beam shaping using SLM, and then the need of their package. But are there other packages (in Python or not) with the same functionality? And if so, what else does pySLM2 bring?

The relative contribution of the three authors is not mentioned. It is assumed that the first author is the main one, but it is necessary to clarify the contributions.
