### 2024-02-27 c38eed8
edit code so that the runtime benchmarking script skip the gpu part is none gpu are found
