+++ runtime_benchmark/runtime_ifta.py
@@ -38,6 +38,7 @@ def task(method):
 os.environ['CUDA_VISIBLE_DEVICES'] = '-1'
 print("Num GPUs Available: ", len(tf.config.list_physical_devices('GPU')))
 # Run the task on CPU
+print("Running on CPU")
 start_time = time.time()
 task('ifta')
 cpu_time = time.time() - start_time
@@ -46,12 +47,16 @@ print(f"Time taken on CPU: {cpu_time} seconds.")
 # Re-enable GPU for the second run (if available)
 os.environ['CUDA_VISIBLE_DEVICES'] = '0'
 
-print("Is Built with CUDA: ", tf.test.is_built_with_cuda())
-print("CUDA Version: ", tf.sysconfig.get_build_info()["cuda_version"])
-print("Num GPUs Available: ", len(tf.config.list_physical_devices('GPU')))
-
-# Run the task on GPU
-start_time = time.time()
-task('ifta')
-gpu_time = time.time() - start_time
-print(f"Time taken on GPU: {gpu_time} seconds.")
+num_gpu = len(tf.config.list_physical_devices('GPU'))
+if num_gpu == 0:
+  print("No GPU found. Exiting.")
+  exit()
+else:
+    print("Num GPUs Available: ", len(tf.config.list_physical_devices('GPU')))
+    print("Is Built with CUDA: ", tf.test.is_built_with_cuda())
+    print("CUDA Version: ", tf.sysconfig.get_build_info()["cuda_version"])
+    # Run the task on GPU
+    start_time = time.time()
+    task('ifta')
+    gpu_time = time.time() - start_time
+    print(f"Time taken on GPU: {gpu_time} seconds.")
