### 2024-02-27 dc71bb0
add some prints
