+++ runtime_benchmark/runtime_ifta.py
@@ -56,6 +56,7 @@ else:
     print("Is Built with CUDA: ", tf.test.is_built_with_cuda())
     print("CUDA Version: ", tf.sysconfig.get_build_info()["cuda_version"])
     # Run the task on GPU
+    print("Running on GPU")
     start_time = time.time()
     task('ifta')
     gpu_time = time.time() - start_time
