### 2024-03-04 QITI/pySLM2#19 @wavefrontshaping
Python version requirements?
This issue is part of my JOSS review at https://github.com/openjournals/joss-reviews/issues/6315

I did not find any Python version requirements or compatibility list, so I tried different versions.

I tried to install requirements with `pip install .` using a fresh python 3.10 conda environment, I got the following errors:

```bash
ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
tts 0.11.1 requires flask, which is not installed.
tts 0.11.1 requires tqdm, which is not installed.
gruut 2.2.3 requires Babel<3.0.0,>=2.8.0, which is not installed.
ipykernel 6.15.3 requires jupyter-client>=6.1.12, which is not installed.
ipykernel 6.15.3 requires psutil, which is not installed.
ipykernel 6.15.3 requires pyzmq>=17, which is not installed.
ipykernel 6.15.3 requires traitlets>=5.1.0, which is not installed.
librosa 0.8.0 requires decorator>=3.0.0, which is not installed.
pandas 1.5.3 requires pytz>=2020.1, which is not installed.
trainer 0.0.20 requires psutil, which is not installed.
tts 0.11.1 requires numpy==1.22.4; python_version == "3.10", but you have numpy 1.26.4 which is incompatible.
numba 0.55.2 requires numpy<1.23,>=1.18, but you have numpy 1.26.4 which is incompatible.
tensorboardx 2.6 requires protobuf<4,>=3.8.0, but you have protobuf 4.25.3 which is incompatible.
trainer 0.0.20 requires protobuf<3.20,>=3.9.2, but you have protobuf 4.25.3 which is incompatible.
```

Similarly, for the docs requirements with `pip install -r docs/requirements.txt`, I got:

```bash
ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
ipython 8.5.0 requires decorator, which is not installed.
ipython 8.5.0 requires jedi>=0.16, which is not installed.
ipython 8.5.0 requires pexpect>4.3; sys_platform != "win32", which is not installed.
ipython 8.5.0 requires prompt-toolkit<3.1.0,>3.0.1, which is not installed.
ipython 8.5.0 requires traitlets>=5, which is not installed.
```

I had no error with a fresh 3.11 env.

If it needs Python>=3.11, that is fine, but it should be stated somewhere.
