### 2024-03-04 QITI/pySLM2#20 @wavefrontshaping
Luxbeam missing in docs for utils
This issue is part of my JOSS review at https://github.com/openjournals/joss-reviews/issues/6315

In the built docs as well as in the online docs, the `utils` section does not mention the `Luxbeam` controller that is used in one example.
