### 2024-03-04 QITI/pySLM2#21 @wavefrontshaping
Control of LCOS-SLMs?
This issue is part of my JOSS review at https://github.com/openjournals/joss-reviews/issues/6315

The description of the package read:
`pySLM2 is a python package for full stack control of using spatial light modulators (SLMs) for holographic beam shaping.`
However, while I did find examples and docs about how to send images to a DMD, I did not when using an LCOS-SLM.

Is it a feature that this package offers? or does it need an external module?
If the features for the DMDs and LOCS-SLMs are different, it would be better to state it.

If it does require another lib to control LCOS-SLMs (with standard monitor connections), a small example on how to do so or a link to an appropriate package would be appreciated.
If so, and I am more than biased here because I wrote it, but [SLMpy](https://github.com/wavefrontshaping/slmPy) could be a good candidate for that.
