### 2024-03-04 QITI/pySLM2#22 @wavefrontshaping
`as_complex()` confusing
This issue is part of my JOSS review at https://github.com/openjournals/joss-reviews/issues/6315

On a quite minor note, I find the function name `as_complex()` for the profiles classes a bit confusing.
Indeed, in numpy/scipy, `as_complex()` only cast an array to a one with a different dtype.
In this case, you create a complex array from a real one by taking the complex exponential of it.
So you convert a phase array to a complex one, which is fundamentally different.
A different name, even slightly different, would limit confusion for the user that would probably expect this function to behave as its namesake in numpy.
