### 2024-03-04 QITI/pySLM2#22 @ldes89150
Agree. Thanks for the suggestion! I will make the changes.
