### 2024-03-04 QITI/pySLM2#19 @ldes89150
Looks like it's the problem of some existing packages in your environment?

We did the testing with multiple python version:
https://github.com/QITI/pySLM2/blob/master/.github/workflows/pytest.yml#L17
