### 2024-03-07 QITI/pySLM2#20 @ldes89150
The API for the Luxbeam controller is here:

https://pyslm2.pages.dev/pySLM2.util#pySLM2.util.controller.LuxbeamController

I will link the example to the doc building process.
