### 2024-03-07 8a3e1b2
gs and ifta runtime benchmarking script, cpu and gpu separated
