### 2024-03-08 05feba2
delete redundant lines in check_gpu
