+++ pySLM2/runtime_benchmark/check_gpu.py
@@ -1,12 +1,6 @@
-# import tensorflow as tf
-
-# # Disable all GPUs
-# tf.config.set_visible_devices([], 'GPU')
-
-# print("Num GPUs Available: ", len(tf.config.list_physical_devices('GPU')))
 import os
 print("CUDA_VISIBLE_DEVICES before TensorFlow import:", os.environ.get('CUDA_VISIBLE_DEVICES'))
-os.environ['CUDA_VISIBLE_DEVICES'] = '-1'
+os.environ['CUDA_VISIBLE_DEVICES'] = '0'
 
 
 import tensorflow as tf
