### 2024-03-08 QITI/pySLM2#19 @wavefrontshaping
> Looks like it's the problem of some existing packages in your environment?
Nah, I used a fresh env using conda

```bash
conda create -n test_pyslm2 python=3.10
```
I tried again, same errors.
Maye some pre-installed packages with conda messes up things?
What tool did you use to create your envs?
