### 2024-03-08 QITI/pySLM2#19 @wavefrontshaping
OK, it seems that the version of pip installed with python 3.10 in conda is buggy and not up to date.
Updating pip after creating the env did the trick, my bad, the error was indeed on my side.
