### 2024-03-10 QITI/pySLM2#13 @brandondube
It is a bit tricky to assess the paper's statement for accuracy, as-is then.  If your iterative transform algorithms do show a speedup but other code runs slower, it is in the weeds of the statement whether it is true because of the "such as" clause, or not true in general because some code is several times slower as-is on GPU.

Looking at your benchmark with 200 iterations, I get a significant non-determinism on the GPU timing; sometimes about 2.2 seconds, other times as long as 3.8 seconds.  In any event, both of these are significantly shorter than the CPU timing. and it is a shared environment.

----

To have a highly interchangeable backend, you could borrow my [BackendShim](https://github.com/brandondube/prysm/blob/master/prysm/mathops.py#L8-L28), which places a tiny bit of indirection in front of "numpy and scipy."  Any library that has the same API can be placed behind the curtain, and the code will work with no rewrite.  You can/should combine this with a [configuration object](https://github.com/brandondube/prysm/blob/master/prysm/conf.py#L5) that controls the numerical precision used.

In our ["Exascale"](https://opg.optica.org/josaa/fulltext.cfm?uri=josaa-39-12-C133&id=521902) paper we reported comparative benchmarks of the same code on 2x Intel 6248R CPUs vs one Titan XP GPU.  Each 6248R has 768 Gflops per Intel, so best case we would expect (12.15/(2*.768)) ~= 8x speedup.  We do a bit better, a bit over 10x, because not all of the algorithms use all CPU cores, while all of them use all GPU cores.  This is the regime you should be in, for code well written to use either CPU or GPU.

I have experimented with most of the machine learning libraries, and found none of them to compete well with CuPy for optics-y code.  Jax can be very picky about all array sizes being known a priori, and both pytorch and tensorflow can get angry if you do certain operations such as array insertion/deletion (which is often used e.g. to zero-pad and crop).  Jax's issues have to do with its JIT compiler being critically dependent on pre-allocating all of the memory for the entire program, and for torch and tf it has to do with their gradient backpropagation logic forbidding some operations such as in-place array modification.

If you do try jax, you will find that it runs most of its functions on about 2 CPU cores (regardless of machine size, from 8 to 96 logical CPUs it is always two).  This makes it the fastest on CPU, in general.  The JIT compiler is mostly useless for optics code, which consists of a sequence of large array operations that cannot be done any more optimally than just writing the code out.  The things that make optics code slow (re-generating the same grid over and over, say) are not something the JIT can elide.

---

For your problems that go slower on GPU than CPU, it is most likely the performance semantics differing between CPU and GPU, and "bad" operations being run on the GPU.  The time for the CPU to do anything is more or less one clock cycle, which is <= 1 ns.  The time for the GPU to do anything is one communication latency over the PCIe bus, which is about 10 microseconds.  Scalar-ish calculations are faster on CPU, but large array calculations are faster on GPU.  In somewhat rare pockets, you will see my code do
```
from prysm.mathops import np

import numpy as truenp

# some operations use truenp
```
This is to run the few things (like spectral weights, say) that would be slower on GPU, on the CPU.

Due to the MIT license, you can copy BackendShim directly into your library, if you like, as long as you copy its LICENSE with it.  If I were re-writing it today I would make config.precision directly take a numpy-esque dtype, instead of an integer number of bits.  Newer GPUs can use the `tensorfloat32` datatype, and be ~ 10x faster with that than IEEE float32.  When the worse rounding is not detrimental to the answer, it is a great-to-have.  As-is my config type precludes using tf32.  You can always sniff the dtype when it is set to find its complex complement with Numpy's more recent improvements to the dtype system.  Likewise for `float16`, although there are very few optics calculations that can really use that dtype without catastrophic rounding.
