### 2024-03-22 80fe165
readme in benchmarking folder. add tensordlow installation instruction
