### 2024-03-22 261c1e3
add loop to runtime tests to report mean, std
