+++ pySLM2/runtime_benchmark/runtime_ifta_gpu.py
@@ -37,9 +37,10 @@ def task(method):
         **kwargs
     )
     end = time.time()
-    print("time used:", end-start)
+    total_time = end-start
+    return total_time
 
-    
+
 
 num_gpu  = len(tf.config.list_physical_devices('GPU'))
 print("Num GPUs Available: ", num_gpu)
@@ -49,8 +50,14 @@ if num_gpu ==0:
 else:
     print("Is Built with CUDA: ", tf.test.is_built_with_cuda())
 
-# Re-enable GPU for the second run (if available)
-task('ifta')
+num_test = 10
+result = []
+print(f'Total {num_test}  Tests Running on GPU')
+for i in range(num_test):
+    ti = task('ifta')
+    print(f'test {i} runtime: {ti:0.02f}s')
+    result.append(ti)
+print(f'runtime for {num_test} runs: {np.mean(result):0.2f}s +- {np.std(result):0.2f}s')
 
 
 
