+++ pySLM2/runtime_benchmark/runtime_ifta_gpu.py
@@ -52,7 +52,7 @@ else:
 
 num_test = 10
 result = []
-print(f'Total {num_test}  Tests Running on GPU')
+print(f'Total {num_test} Tests Running on GPU')
 for i in range(num_test):
     ti = task('ifta')
     print(f'test {i} runtime: {ti:0.02f}s')
