### 2024-03-23 7c91f42
table for runtime, reported mean and std
