### 2024-03-27 6f6090e
add mraf tests. add all results in readme and added cpu info
