+++ pySLM2/runtime_benchmark/slm_gs_cpu.py
@@ -8,36 +8,30 @@ import pySLM2
 import tensorflow as tf
 
 
-def task(method):
-    dmd = pySLM2.DLP9500(
-        369 * nano,  # wavelength
-        200 * milli, # effective focal length
-        4, # periodicity of the grating = 4 pixels
-        np.pi/4, # The dmd is rotated by 45 degrees
-    )
-    x0, y0 = dmd.first_order_origin
+def task(method, N):
+    lcos_slm = pySLM2.PLUTO_2(
+    369 * nano,  # wavelength
+    200 * milli, # effective focal length
+  )
+
     # The beam illumilating the DMD is an gaussian beam with a waist of 5 mm
     input_profile = pySLM2.HermiteGaussian(0,0,1,5*milli)
 
     # targeted profile at the image plane
-    # Here we create two gaussian spots separated by 30 microns
-    output_profile = pySLM2.HermiteGaussian(0,0,1,10*micro, n=0, m=0) - pySLM2.HermiteGaussian(30*micro,0,1,10*micro, n=0, m=0)
-    signal_window = pySLM2.RectangularWindowRectangle(x0, y0, 100 * micro, 100 * micro)
+    output_profile = pySLM2.HermiteGaussian(0,0,1,30*micro, n=1, m=1)
 
     start = time.time()
-
-    kwargs = dict()
-    kwargs["signal_window"] = signal_window
-    kwargs["N"] = 2000
-    dmd.calculate_dmd_state(
+    lcos_slm.calculate_hologram(
         input_profile,
         output_profile,
         method=method,
-        **kwargs
+        N=N,
     )
     end = time.time()
     total_time = end-start
     return total_time
+
+
 
 num_gpu  = len(tf.config.list_physical_devices('GPU'))
 print("Num GPUs Available: ", num_gpu)
@@ -47,15 +41,17 @@ if num_gpu ==0:
 else:
     print("Found GPU. Exit.")
     exit()
-# Run the task on CPU
-print("Running on CPU")
 
+N=1000
 num_test = 10
 result = []
+
 print(f'Total {num_test} Tests Running on CPU')
 for i in range(num_test):
-    ti = task('ifta')
+    ti = task('gs',N)
     print(f'test {i} runtime: {ti:0.02f}s')
     result.append(ti)
 print(f'runtime for {num_test} runs: {np.mean(result):0.2f}s +- {np.std(result):0.2f}s')
+
+
 
