### 2024-03-28 3b89038
add gpu doc in root readme
