+++ README.md
@@ -23,7 +23,7 @@ The dependencies of `pySLM2` in includes: `numpy`, `scipy`, `matplotlib`, and `t
 Installation
 ------------
 
-### With setuptools
+### pySLM2 installation with setuptools
 
 If you prefer the development version from GitHub, download it here, `cd` to the pySLM2 directory, and use:
 ```
@@ -36,8 +36,9 @@ Or, if you wish to edit the pySLM2 source code without re-installing each time
 pip install -e .
 ```
 
-Tensorflow Installation for GPU Support
+GPU Support via Tensorflow
 ---------------------------------------
+### Tensorflow installation tips
 
 `pySLM2` primarily relies on `tensorflow` for most of its numerical computations. For machines with compatible hardware, `tensorflow` can seamlessly utilize GPU acceleration to enhance performance, provided it is installed correctly.
 
@@ -47,11 +48,31 @@ The exact package dependencies vary depending on each system configuration and t
  As a reference, in our setup with Windows 10 Build 17763 and an NVIDIA Quadro M4000 GPU, we executed the following commands to install `tensorflow`-related packages in a Conda environment with `python=3.7`.
 ```
 conda install -c conda-forge cudnn==7.6.5
-conda install -c conda-forge cudatoolkit=10.1.243
+conda install -c conda-forge cudatoolkit==10.1.243
  
 pip install tensorflow==2.1 
 pip install tensorflow-gpu==2.1 
 ```
+
+### Runtime Benckmarking: CPU vs GPU 
+Several runtime benchmarking scripts for iterative hologram generations algorithms are included in a separate foler `pySLM2/runtime_benchmark`. Instructions for running those tests can be found in [`pySLM2/runtime_benchmark/README.md`](pySLM2/runtime_benchmark/README.md).
+
+#### Runtime Benckmarking Example
+##### Algorithm performance comparision: Intel Core i9-9900K CPU vs NVidia Quadro M4000 GPU
+Our machine has Windows 10 Build 17763 with Intel Core i9-9900K CPU and an NVidia Quadro M4000 GPU. Key package dependencies are:
+- `python`: 3.7.1
+- `tensorflow`: 2.1.0
+- `tensorflow-gpu`: 2.1.0
+- `cudnn`: 7.6.5
+- `cudatookkit`: 10.1.243
+
+From our testing, we observed:
+|  | Case 1 (gs)   | Case 2 (mraf)  | Case 3 (ifta)
+|-------------|-------------|-------------|-------------|
+| CPU |$225.11 \pm 3.98$ s | $221.64 \pm 1.78$ s  | $206.44 \pm 0.49$ s|
+| GPU  | $6.48 \pm 0.29 $ s | $8.29 \pm 0.31$ s |$6.76 \pm 0.50$ s|
+
+These findings show that the iterative algorithms can be greatly accelerated by GPU usage.
 
 Optional Dependencies for Hardware Controls
 -------------------------------------------
