### 2024-04-07 1a6864e
13 installation gpu acceleration (#17)
* move documentation building instructions in docs/README.md. Adding dependencies section to README.md in root folder

* fix typo

* more detailes on dependencies and tensorflow installation

* add runtime benchmarking scripts

* fix typo

* edit code so that the runtime benchmarking script skip the gpu part is none gpu are found

* add some prints

* put runtime_benchmark folder in pySLM2

* gs and ifta runtime benchmarking script, cpu and gpu separated

* delete redundant lines in check_gpu

* readme for runtime benckmarking. results part

* readme in b
[…]
