+++ paper.md
@@ -99,7 +99,7 @@ One of the goals of `pySLM2` is to abstract the hardware details and offer a uni
 
 
 # Author Contributions
-C.-Y.S. designed and implemented the package. J.Z. contributed to implementing the hardware controls and the performance benchmarking. R.I. advised on the scientific aspects of the package. All authors contributed to the writing of the manuscript and the documentation of the package.
+C.-Y.S. designed and implemented the package. J.Z. contributed to implementing the hardware controls and performance benchmarking. R.I. advised on the scientific aspects of the package. All authors contributed to writing the manuscript and documenting the package.
 
 # Acknowledgements
 The hardware controls for the DMDs from ViALUX GmbH in the package is built on top of the `AL4lib` [@sebastien_m_popoff_2022_6121191]. We appreciate the work of the authors of `AL4lib`. We express gratitude to Kaleb Ruscitti for assisting with hardware testing and to Sainath Motlakunta and Nikhil Kotibhaskar for providing valuable feedback on the package. 
