+++ paper.bib
@@ -173,3 +173,32 @@
   volume={656},
   year={1971}
 }
+
+@article{nishino2017cupy,
+  title={Cupy: A numpy-compatible library for nvidia gpu calculations},
+  author={Nishino, ROYUD and Loomis, Shohei Hido Crissman},
+  journal={31st confernce on neural information processing systems},
+  volume={151},
+  number={7},
+  year={2017}
+}
+
+@inproceedings{yamamoto2021gradient,
+  title={Gradient-based Optimization of Time-Multiplexed Binary Computer-Generated Holograms by Digital Mirror Device},
+  author={Yamamoto, Kenta and Ochiai, Yoichi},
+  booktitle={Digital Holography and Three-Dimensional Imaging},
+  pages={DTh7C--1},
+  year={2021},
+  organization={Optica Publishing Group}
+}
+
+@article{matthes2019optical,
+  title={Optical complex media as universal reconfigurable linear operators},
+  author={Matth{\`e}s, Maxime W and Del Hougne, Philipp and De Rosny, Julien and Lerosey, Geoffroy and Popoff, S{\'e}bastien M},
+  journal={Optica},
+  volume={6},
+  number={4},
+  pages={465--472},
+  year={2019},
+  publisher={Optica Publishing Group}
+}
