### 2024-04-07 323a013
add opensource options for lcos
