### 2024-04-08 QITI/pySLM2#13 @brandondube
I wanted to let you know that while you have closed this issue, you have not resolved the whole of the problem.  The documentation and papers' performance claims must reflect accurately the software's performance.  As-is, broad claims are made about the code running faster on a GPU, while only a minority fraction of the library's scope does.
