### 2024-04-08 QITI/pySLM2#13 @ldes89150
GitHub automatically closed the issue after merging the PR link to it. Let's me reopen it.
