### 2024-04-08 QITI/pySLM2#13 @ldes89150
> I added tf.debugging.set_log_device_placement(True) to the collab notebook and ran examples/create_multiple_gaussian_beam.py there. It took ~16 sec, and tf indicated everything ran on the GPU. It takes about two seconds on my machine, with CPU. Understanding that the collab environment is quite different to my machine which is bare metal, I disabled the GPU in collab with tf.config.set_visible_devices([], 'GPU'). The debug printing I added showed everything was being done on CPU, and the sample example took 7 seconds. These timings are repeatable.

In this example, the majority of the time is actually used in plotting. (To make it faster, I should trim the array before feeding the data into `pcolormesh` for plotting). The computation for both CPU & GPU part takes <1 seconds. Comparing to the IO overhead, the computational speed up is not significant for this algorithm. The algorithms benefits the most are those relying on iterative Fourier transformations. They are also the ones have much higher computational costs. I will mention whether the algorithms benefit from the GPU in the document for better clarity.

As for JIT, I agree that JIT doesn't always help, but I do observe speed up with iterative Fourier transformation type algorithms.  I think what JIT helps here is reducing the memory allocation/deallocation for the intermediate results. (JIT was an experimental features in some earlier tensorflow versions. It's switch was in the experimental configs in the earlier versions.)

Here I put some code to show the performance difference with/without  JIT (`set_jit`) and with/without tracing (`run_functions_eagerly`):

https://colab.research.google.com/gist/ldes89150/73bc961960c79a730660e966a035c2aa/pyslm2.ipynb

This is the result running on a T4 instance (GPU):
```
No JIT With Tracing (Default)
time used: 3.390634536743164
time used: 3.3433678150177
time used: 3.1825449466705322
time used: 3.186124324798584
time used: 3.351879596710205
Enable JIT
time used: 2.9088780879974365
time used: 2.7242681980133057
time used: 2.7510483264923096
time used: 2.86055850982666
time used: 2.881774425506592
No JIT No Tracing (Without Optimization)
time used: 3.533553123474121
time used: 3.539891481399536
time used: 3.7536377906799316
time used: 3.537489652633667
time used: 3.535156726837158
```
