### 2024-04-08 QITI/pySLM2#13 @brandondube
It is difficult for me to accept on face value your claim that the computation is less than one second when running the same code on CPU and GPU takes 7 and 16 seconds, respectively, when those timings are repeatable.

In any event, the claim in the paper and documentation is that

> pySLM2 primarily relies on tensorflow for most of its numerical computations. For machines with compatible hardware, tensorflow can seamlessly utilize GPU acceleration to enhance performance, provided it is installed correctly.

Some algorithms running twice as slowly on GPU does not match this claim
