### 2024-04-08 QITI/pySLM2#13 @ldes89150
To address your concern, I will update the claims for better clarity in both the document and the paper to emphasize the GPU acceleration is for the computational heavy tasks.
