### 2024-04-08 25c5dd5
adding dedicated algorithm API. docstring and detailed explanation in progess.
