### 2024-04-08 10926ca
docstring for algorithms in progress
