### 2024-04-09 5eeaf3a
docstring in slm.py
