### 2024-04-11 4178b00
contents for algorithms explanation on sphinx. docstring for some complicated functions in _lib.py and slm.py are added
