+++ README.md
@@ -10,11 +10,12 @@ pySLM2
 
 * API Docs: https://pyslm2.pages.dev/
 
-Instructions to build documentation locally can be found in `docs/README.md`.
+Instructions to build documentation locally can be found in [`docs/README.md`](docs/README.md).
 
 
 Dependencies
 ------------
+pySLM2 supports Python 3.9+. 
 
 The dependencies of `pySLM2` in includes: `numpy`, `scipy`, `matplotlib`, and `tensorflow`.
 
@@ -39,7 +40,7 @@ GPU Support via Tensorflow
 ---------------------------------------
 ### Tensorflow installation tips
 
-`pySLM2` primarily relies on tensorflow for most of its numerical computations. For machines with compatible hardware, tensorflow can seamlessly utilize GPU acceleration to enhance the performance of some numerically heavy tasks, provided it is installed correctly. Algorithms or operations that heavily rely on numerical Fourier transformation especially benefit from GPU acceleration.
+`pySLM2` primarily relies on `tensorflow` for most of its numerical computations. For machines with compatible hardware, `tensorflow` can seamlessly utilize GPU acceleration to enhance performance, provided it is installed correctly.
 
 The exact package dependencies vary depending on each system configuration and the GPU card. For details about machine compatibility and correct version of tensorflow, please refer to the [Tensorflow's website](https://www.tensorflow.org/install/pip#step-by-step_instructions), which provides installation guide for different operating systems. Another authors' recommended `tensorflow` installation guide can also be found in this [website](https://medium.com/@shaikhmuhammad/installing-tensorflow-cuda-cudnn-with-anaconda-for-geforce-gtx-1050-ti-79c1eb94eb7a) which provides thorough information about package dependecies such as `cuda` and `cudnn` versions. 
 
@@ -54,7 +55,7 @@ pip install tensorflow-gpu==2.1
 ```
 
 ### Runtime Benckmarking: CPU vs GPU 
-Several runtime benchmarking scripts for iterative hologram generations algorithms are included in a separate foler `pySLM2/runtime_benchmark`. Instructions for running those tests can be found in `pySLM2/runtime_benchmark/README.md`.
+Several runtime benchmarking scripts for iterative hologram generations algorithms are included in a separate foler `pySLM2/runtime_benchmark`. Instructions for running those tests can be found in [`pySLM2/runtime_benchmark/README.md`](pySLM2/runtime_benchmark/README.md).
 
 #### Runtime Benckmarking Example
 ##### Algorithm performance comparision: Intel Core i9-9900K CPU vs NVidia Quadro M4000 GPU
@@ -66,11 +67,10 @@ Our machine has Windows 10 Build 17763 with Intel Core i9-9900K CPU and an NVidi
 - `cudatookkit`: 10.1.243
 
 From our testing, we observed:
-
 |  | Case 1 (gs)   | Case 2 (mraf)  | Case 3 (ifta)
 |-------------|-------------|-------------|-------------|
 | CPU |$225.11 \pm 3.98$ s | $221.64 \pm 1.78$ s  | $206.44 \pm 0.49$ s|
-| GPU  | $6.48 \pm 0.29$ s | $8.29 \pm 0.31$ s |$6.76 \pm 0.50$ s|
+| GPU  | $6.48 \pm 0.29 $ s | $8.29 \pm 0.31$ s |$6.76 \pm 0.50$ s|
 
 These findings show that the iterative algorithms can be greatly accelerated by GPU usage.
 
