### 2024-05-22 3ad74c4
add author contribution (#23)
* add author contribution

* update author contribution

* update

* update

* add opensource options for lcos
