### 2024-05-22 QITI/pySLM2#18 @ldes89150
Addressed in the following two issues
- Author contribution: #23
- Statement of need update:  #25
