### 2024-05-27 98cceec
improved algo description. added references. need to check if references are correct
