+++ docs/ref.bib
@@ -1,0 +1,175 @@
+@article{shih2021reprogrammable,
+  title={Reprogrammable and high-precision holographic optical addressing of trapped ions for scalable quantum control},
+  author={Shih, Chung-You and Motlakunta, Sainath and Kotibhaskar, Nikhil and Sajjan, Manas and Habl{\"u}tzel, Roland and Islam, Rajibul},
+  journal={npj Quantum Information},
+  volume={7},
+  number={1},
+  pages={57},
+  year={2021},
+  publisher={Nature Publishing Group UK London}
+}
+
+@article{zupancic2016ultra,
+  title={Ultra-precise holographic beam shaping for microscopic quantum control},
+  author={Zupancic, Philip and Preiss, Philipp M and Ma, Ruichao and Lukin, Alexander and Tai, M Eric and Rispoli, Matthew and Islam, Rajibul and Greiner, Markus},
+  journal={Optics express},
+  volume={24},
+  number={13},
+  pages={13881--13893},
+  year={2016},
+  publisher={Optica Publishing Group}
+}
+
+@incollection{lee1978iii,
+  title={III computer-generated holograms: Techniques and applications},
+  author={Lee, Wai-Hon},
+  booktitle={Progress in optics},
+  volume={16},
+  pages={119--232},
+  year={1978},
+  publisher={Elsevier}
+}
+
+@article{gerhberg1972practical,
+  title={A practical algorithm for the determination of phase from image and diffraction plane picture},
+  author={Gerhberg, R and Saxton, W},
+  journal={Optik},
+  volume={35},
+  pages={237--246},
+  year={1972}
+}
+
+@article{motlakunta2023preserving,
+  title={Preserving a qubit during adjacent measurements at a few micrometers distance},
+  author={Motlakunta, Sainath and Kotibhaskar, Nikhil and Shih, Chung-You and Vogliano, Anthony and Mclaren, Darian and Hahn, Lewis and Zhu, Jingwen and Habl{\"u}tzel, Roland and Islam, Rajibul},
+  journal={arXiv preprint arXiv:2306.03075},
+  year={2023}
+}
+
+@software{sebastien_m_popoff_2022_6121191,
+  author       = {Sébastien M. Popoff and
+                  Gilbert Shih and
+                  Dirk B. and
+                  GustavePariente},
+  title        = {wavefrontshaping/ALP4lib: 1.0.1},
+  month        = feb,
+  year         = 2022,
+  publisher    = {Zenodo},
+  version      = {1.0.1},
+  doi          = {10.5281/zenodo.6121191},
+  url          = {https://doi.org/10.5281/zenodo.6121191}
+}
+
+@article{gaunt2012robust,
+  title={Robust digital holography for ultracold atom trapping},
+  author={Gaunt, Alexander L and Hadzibabic, Zoran},
+  journal={Scientific reports},
+  volume={2},
+  number={1},
+  pages={721},
+  year={2012},
+  publisher={Nature Publishing Group UK London}
+}
+
+@article{qian2021super,
+  title={Super-resolved imaging of a single cold atom on a nanosecond timescale},
+  author={Qian, Zhong-Hua and Cui, Jin-Ming and Luo, Xi-Wang and Zheng, Yong-Xiang and Huang, Yun-Feng and Ai, Ming-Zhong and He, Ran and Li, Chuan-Feng and Guo, Guang-Can},
+  journal={Physical review letters},
+  volume={127},
+  number={26},
+  pages={263603},
+  year={2021},
+  publisher={APS}
+}
+
+@article{drechsler2021optical,
+  title={Optical superresolution sensing of a trapped ion’s wave packet size},
[…]
