+++ docs/md_alg.md
@@ -1,0 +1,50 @@
+# Fourier Hologram Generation Algorithms
+
+## Overview
+
+pySLM2 provides a suite of hologram generation algorithms optimized for various types of Spatial Light Modulator (SLM) modules, tailored to their specific phase and amplitude modulation capabilities.
+
+For Liquid Crystal on Silicon (LCoS) modules, which are capable of pure phase modulation, we offer the Gerchberg-Saxton (`gs`) and Mixed-Region Amplitude Freedom (`mraf`) algorithms. [[3]](#ref3)
+
+For SLMs, use the method keyword argument in the `slm.calculate_hologram` function to apply the desired algorithm.
+
+For Digital Micromirror Devices (DMDs), which enable both amplitude and phase modulation, we implement specialized methods for different control types. The `ideal` method is used for simulating hologram when having continuous amplitude control on DMD. Three binarization methods are packaged along with it: the `simple` method for hard cut-off binatization, `random` method for randomized binarization [[2]](#ref2), and an interative Fourier transformation algorithm `ifta` with adaptive binarization.  
+
+For DMDs, use the method keyword argument in the `slm.calculate_dmd_state` function to appply the desired algorithm.
+
+## Functions
+
+### pySLM2.slm.LCOS_SLM.calculate_hologram()
+
+- `gs`: The Gerchberg-Saxton algorithm is an iterative phase retrieval process. The refined phase profile ideally would converge to the target phase profile after enough iterations. However, this convergence can be slow and is not guaranteed. [[4]](#ref4)
+- `mraf`: The Mixed-Region Amplitude Freedom algorithm is an improved adpataion of the `gs` algorihtms which interatively refines the phase profile within a specified signal window region. It improves the convergence within the signal window at the sacrifice of phase control in the outside region. [[3]](#ref3)
+
+### pySLM2.slm.DMD.calculate_dmd_state()
+
+- `ideal`: It assumes continuous amplitude control, generating the ideal grey-scale hologram with no binarization induced errors.
+- `simple`: It binarizes the ideal hologram with a 0.5 threshold for the grating phase. However, the hard cut-off can induce significant binarization induced errors.
+- `random`: It binarizes the ideal hologram with random threshold by probabilisticly turn on or off pixel mirrors based on the grating phase. It reduces binarization artifacts compared to the `simple` method. [[2]](#ref2).
+- `ifta`: It interatively refines the holgram within a signal window by deterministically binarize the pixel mirror with adaptive binarizaitno threshold. This method yields lower RMS error within the signal window compared to the `random` method. [[1](#ref1), [5](#ref5)]
+
+
+## References
+
+<a id="ref1"></a>1. Chung-You Shih, Sainath Motlakunta, Nikhil Kotibhaskar, Manas Sajjan, Roland Hablützel, Rajibul Islam. "Reprogrammable and high-precision holographic optical addressing of trapped ions for scalable quantum control." *npj Quantum Information*, vol. 7, no. 1, p. 57, 2021. Nature Publishing Group UK London.
+
[…]
