### 2024-05-28 f548c98
update the algorithm introduction. added a few docstring  lines in _lib as well.
