+++ docs/readme_link.rst
@@ -1,1 +1,1 @@
-.. mdinclude:: ../README.md
+.. mdinclude:: README.md
