+++ pySLM2/_lib.py
@@ -39,11 +39,35 @@ def _inverse_fourier_transform(profile_tensor):
 
 
 def calculate_dmd_grating(amp, phase_in, phase_out, x, y, p, theta, method="random", negative_order=False, **kwargs):
+    '''
+    Calculate a DMD grating pattern. Algorithms available are "ideal", "simple", "random" and "ifta".
+    
+    Parameters
+    ----------
+    amp : float
+        Amplitude of the grating pattern.
+    phase_in : float
+        Phase of the grating pattern at the input plane.
+    phase_out : float
+        Phase of the grating pattern at the output plane.
+    x : tf.Tensor
+        X coordinate of the plane.
+    y : tf.Tensor
+        Y coordinate of the plane.
+    p : float
+        Period of the grating pattern.
+    theta : float
+        Angle of the grating pattern.
+    method : str
+        Method to calculate the grating pattern. Options are "ideal", "simple", "random", "ifta".
+    negative_order : bool
+        If True, the grating pattern will be calculated for the negative order.
+    **kwargs : dict
+        Additional parameters for the selected method.
+    '''
     negative_order = tf.constant(negative_order, dtype=tf.bool)
     if method == "ideal":
         return _calculate_dmd_grating_ideal(amp, phase_in, phase_out, x, y, p, theta, negative_order=negative_order)
-    elif method == "ideal_square":
-        return _calculate_dmd_grating_ideal_square(amp, phase_in, phase_out, x, y, p, theta, negative_order=negative_order)
     elif method == "simple":
         return _calculate_dmd_grating_simple(amp, phase_in, phase_out, x, y, p, theta, negative_order=negative_order)
     elif method == "random":
@@ -67,11 +91,57 @@ def calculate_dmd_grating(amp, phase_in, phase_out, x, y, p, theta, method="rand
 
 @tf.function
 def _grating_phase(x, y, theta, p):
+    '''
+    Calculate the phase of a grating pattern.
+    
+    Parameters
+    ----------
+    x : tf.Tensor
+        X coordinate of the plane.
+    y : tf.Tensor
+        Y coordinate of the plane.  
+    theta : float
+        Angle of the grating pattern.
+    p : float
+        Period of the grating pattern.
+
+    Returns
+    -------
+    tf.Tensor
+        Phase of the grating pattern.
+    '''
     return (2 * math.pi / p) * (tf.cos(theta) * x + tf.sin(theta) * y)
 
 
 @tf.function
 def _calculate_dmd_grating_ideal(amp, phase_in, phase_out, x, y, p, theta, negative_order=False):
+    '''
+    Calculate a grey scale DMD grating pattern.
+    
+    Parameters
+    ----------
+    amp : float
+        Amplitude of the grating pattern.
+    phase_in : float
+        Phase of the grating pattern at the input plane.
+    phase_out : float
+        Phase of the grating pattern at the output plane.
+    x : tf.Tensor
+        X coordinate of the plane.
+    y : tf.Tensor
+        Y coordinate of the plane.
+    p : float
+        Period of the grating pattern.
+    theta : float
+        Angle of the grating pattern.
+    negative_order : bool
+        If True, the grating pattern will be calculated for the negative order.
+    
+    Returns
+    -------
+    tf.Tensor
+        Grating pattern.
+    '''
     grating_phase = _grating_phase(x, y, theta, p)
     if negative_order:
         phase_in = -phase_in
@@ -80,18 +150,12 @@ def _calculate_dmd_grating_ideal(amp, phase_in, phase_out, x, y, p, theta, negat
     grating = amp * (tf.cos(grating_phase - (phase_out - phase_in)) / 2 + 0.5)
[…]
