+++ docs/contributing.md
@@ -1,0 +1,68 @@
+# Contributing to [Your Package Name]
+
+Thank you for considering contributing to `pySLM2`! Here are a few guidelines to help you get started.
+
+## Table of Contents
+
+1. [How to Contribute](#how-to-contribute)
+    - [Reporting Issues](#reporting-issues)
+    - [Submitting Code Changes](#submitting-code-changes)
+    - [Improving Documentation](#improving-documentation)
+2. [Support](#support)
+
+## How to Contribute
+
+### Reporting Issues
+
+If you find a bug or have a feature request, please create an issue on our [GitHub Issues page](https://github.com/QITI/pySLM2/issues).
+
+**When reporting an issue, please include:**
+- A clear, descriptive title.
+- A detailed description of the problem.
+- Steps to reproduce the issue.
+- Any relevant logs, screenshots, or examples.
+
+### Submitting Code Changes
+
+We welcome contributions in the form of pull requests (PRs). Here is how you can submit your changes:
+
+1. **Fork the repository**: Click the "Fork" button on the top right corner of the repository page.
+2. **Clone your fork**: 
+    ```bash
+    git clone https://github.com/QITI/pySLM2.git
+    ```
+3. **Create a new branch**: 
+    ```bash
+    git checkout -b your-feature-branch
+    ```
+4. **Make your changes**: Implement your feature, fix the bug, or improve the documentation.
+5. **Commit your changes**: 
+    ```bash
+    git commit -m "Description of the changes"
+    ```
+6. **Push to your fork**: 
+    ```bash
+    git push origin your-feature-branch
+    ```
+7. **Submit a pull request**: Go to the original repository, click the "New Pull Request" button, and follow the instructions.
+
+**Please make sure your pull request:**
+- Follows the coding style and conventions of the project.
+- Includes tests if applicable.
+- Is well-documented.
+
+### Improving Documentation
+
+Documentation improvements are always welcome. If you find any part of the documentation unclear or incomplete, feel free to submit an issue or a pull request with your improvements.
+
+## Support
+
+If you need help with `pySLM2`, you can:
+
+1. Check the [documentation](https://github.com/yourusername/yourpackagename/wiki).
+2. Browse existing issues or create a new one on our [GitHub Issues page](https://github.com/QITI/pySLM2/issues).
+3. Contact the maintainers via email: j436zhu@uwaterloo.ca.
+
+We will do our best to assist you!
+
+Thank you for contributing to `pySLM2`!

[…]
