### 2024-05-30 ea97a14
fix typo in algorithm.md
