### 2024-05-30 a278ab5
update doc building dep in github workflow and requirements
