+++ .github/workflows/sphinx.yml
@@ -28,6 +28,7 @@ jobs:
         python -m pip install build
         pip install .
         python -m pip install matplotlib furo m2r2 sphinx sphinx-math-dollar sphinx_pyproject
+        python -m pip install docutils==0.16
     - name: Build Document
       working-directory: docs/
       run: make html
