### 2024-05-30 4e0e824
Revert "update doc building dep in github workflow and requirements"
This reverts commit a278ab544c702c80c57b39f5499f0e383b2705c0.
