+++ docs/requirements.txt
@@ -4,4 +4,3 @@ m2r2
 sphinx
 sphinx-math-dollar
 sphinx_pyproject
-docutils==0.16
