+++ _sources/requirements.txt
@@ -1,6 +1,6 @@
 matplotlib
 furo
 m2r2
-sphinx
+sphinx==7.2.3
 sphinx-math-dollar
 sphinx_pyproject
