### 2024-05-31 e93f902
link contributing instruction to sphinx
