+++ docs/contributing_link.rst
@@ -1,0 +1,1 @@
+.. mdinclude:: contributing.md
