+++ docs/contributing_link.rst
@@ -1,1 +1,1 @@
-.. mdinclude:: contributing.md
+.. mdinclude:: ../CONTRIBUTING.md
