### 2024-06-02 9efbf68
Deploying to gh-pages from @ QITI/pySLM2@8f79ccc7415f8d50a418937cdca4f6a1eaf34ebf 🚀
