### 2024-06-02 QITI/pySLM2#14 @ldes89150
> Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
> This criteria is present only in your README; please add it to the docs website

`README.md` is automatically included in the document building process now.

The dependencies are also specified in the toml file.
https://github.com/QITI/pySLM2/blob/master/pyproject.toml#L10

> This criteria is present only in your README; please add it to the docs website
>
>  Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
> Please add a CONTRIBUTING.[md/txt] or likewise

Contribution guideline added: https://pyslm2.pages.dev/contributing_link
