### 2024-06-02 QITI/pySLM2#15 @ldes89150
> These seem somewhat contradictory; the first quote implies Lee's hologram algorithm is present in the package, the latter pair imply it is not. I was not able to find it. All I could find is in _lib.py, the def calculate_dmd_grating function and the routines it calls. I was surprised that none of the code in _lib.py indicates which function is which algorithm. Could you please name the functions or add to their documentation which is which?

Documentation for the algorithms: https://pyslm2.pages.dev/algorithm
The Lee hologram corresponds to the `ideal` and the `simple` methods.
