+++ paper.bib
@@ -202,3 +202,14 @@
   year={2019},
   publisher={Optica Publishing Group}
 }
+
+@article{pasienski2008high,
+  title={A high-accuracy algorithm for designing arbitrary holographic atom traps},
+  author={Pasienski, Matthew and DeMarco, Brian},
+  journal={Optics express},
+  volume={16},
+  number={3},
+  pages={2176--2190},
+  year={2008},
+  publisher={Optica Publishing Group}
+}
