### 2024-06-02 QITI/pySLM2#15 @ldes89150
> Unrelatedly, a grammar issue: ...improved predecessors is improper English. To be a predecessor, Shih et al or Zupancic et al would have had to publish their algorithms before Lee. These could perhaps be considered "successors" but are more aptly "improved alternatives."

Fixed!
