### 2024-06-02 QITI/pySLM2#16 @ldes89150
>  A statement of need: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?
>
> Missing ...and its relation to other work?
>
>  State of the field: Do the authors describe how this software compares to other commonly-used packages?
> no

Addressed in #25
