### 2024-06-02 QITI/pySLM2#20 @ldes89150
Example for the `LuxbeamController`

https://pyslm2.pages.dev/examples#load-image-to-the-luxbeam-dmd-controller
