+++ _downloads/e83f39f258f53ba396bab875f8c1b160/aberration_correction.py
@@ -26,7 +26,7 @@ aberration = pySLM2.Zernike(10, 5 * milli, n=4, m=0)
 
 # The beam illumilating the DMD is an gaussian beam with a waist of 5 mm
 input_profile_unaware_of_aberration = pySLM2.HermiteGaussian(0,0,1,5*milli)
-input_profile = input_profile_unaware_of_aberration  * aberration.as_complex()
+input_profile = input_profile_unaware_of_aberration  * aberration.as_complex_profile()
 
 # targeted profile at the image plane
 output_profile = pySLM2.HermiteGaussian(0,0,1,10*micro, n=0, m=0)
