### 2024-06-06 QITI/pySLM2#22 @ldes89150
Addressed in https://github.com/QITI/pySLM2/commit/dc62bf207b86ebb87ec54a91977a0017392dbfe9
