+++ README.md
@@ -95,6 +95,10 @@ The DMD from Visitech are communicates with UDP, and therefore no driver is need
 pip install Luxbeam
 ```
 
+### Notes for LCOS-SLMs
+
+For LCoS-SLMs, most models can be directly controlled via standard monitor connections. To display holograms, one might consider using slmPy. Since pySLM2 calculates hologram values in radians, one will need a conversion table to determine the corresponding grayscale values for output. This conversion typically varies based on the light wavelength and the specific model of the LCoS-SLM.
+
 Floating Point Precision
 ------------------------
 Since not all the graphic cards are optimized for double-precision computation, the default precision of pySLM2 is set to single-precision.
