### 2024-06-06 QITI/pySLM2#21 @ldes89150
Included in the document!
https://github.com/QITI/pySLM2/blob/master/README.md#notes-for-lcos-slms
