### 2024-06-06 #6315 @ldes89150
Hey @HaoZeke @wavefrontshaping @brandondube @sidihamady

I believe I have addressed all the issues related to the documents and the code mentioned in the [issues](https://github.com/QITI/pySLM2/issues). Please take a look!

I still plan to make some changes to the main text of the paper, but this should be completed in a few days.
Thank you so much for your time! I really appreciate it.
