### 2024-06-06 QITI/pySLM2#21 @wavefrontshaping
Good, thanks! Could you also add the link to the slmPy repo in the doc? (https://github.com/wavefrontshaping/slmPy)
