### 2024-06-22 e2dd766
rearrange statement of need paragraph. fixed BEC typo
