+++ paper.md
@@ -43,15 +43,14 @@ For hardware control, pySLM2 offers a unified application interface (API) compat
 # Statement of need
 High-quality optical controls are crucial for numerous scientific and engineering applications. For instance, in atom-based quantum information processors, quantum states of individual atoms are often manipulated by individually addressing laser beams. The quality of these addressing beams directly impacts the fidelity of quantum operations [@motlakunta2023preserving].
 
-Holographic beam shaping using SLMs provides a way for precise and adaptive optical controls. Compared to using conventional optical elements, holographic beam shaping has several advantages. Firstly, it can generate arbitrary beam profiles that are challenging to create with standard optical elements. For example, the Laguerre-Gaussian beam with a non-zero azimuthal index (often referred to as a doughnut beam), which can be used to trap atoms in a tube-like potential [@kuga1997novel], apply angular momentum to Bose-Einstein Condensation [@andersen2006quantized], or achieve super-resolution imaging [@qian2021super;@drechsler2021optical].
+Holographic beam shaping using SLMs provides a way for precise and adaptive optical controls. Compared to using conventional optical elements, holographic beam shaping has several advantages. Firstly, it can generate arbitrary beam profiles that are challenging to create with standard optical elements. For example, the Laguerre-Gaussian beam with a non-zero azimuthal index (often referred to as a doughnut beam), which can be used to trap atoms in a tube-like potential [@kuga1997novel], apply angular momentum to Bose-Einstein Condensate [@andersen2006quantized], or achieve super-resolution imaging [@qian2021super;@drechsler2021optical].
 
 
 Secondly, holographic beam shaping can actively correct optical aberrations in the system, thereby achieving diffraction-limited performance. This enables the faithful production of target beam profiles with high accuracy. It has been shown that residual aberrations can be corrected to less than $\lambda/20$ root-mean-square (RMS) [@shih2021reprogrammable;@zupancic2016ultra], which is smaller than most of the manufacturing tolerances of optical elements.
 
 At the time of writing, the `pySLM2` package, as detailed in this manuscript, has been used in the trapped ion quantum information processing research [@shih2021reprogrammable;@motlakunta2023preserving;@kotibhaskar2023programmable]. The authors believe that the package will benefit a broader community of researchers and engineers by offering turnkey solutions for applying holographic beam shaping to their work. Moreover, the primitives included in the package can assist researchers in rapidly prototyping new hologram generation algorithms.
 
-Comparing to the exising open-source projects, `pySLM2` provides more inclusive support for binay-amplitude holograms suitable for DMDs. On the other hand, [`slmsuite`](https://slmsuite.readthedocs.io/) provides a comprehensive suite of hologram algorithms tailored for phase-only Spatial Light Modulators (SLMs), featuring GPU acceleration via `CuPy` [@nishino2017cupy]. Additionally, `slmsuite` includes hardware control capabilities for Liquid Crystal on Silicon (LCoS) SLMs. [`CGH-diff`](https://github.com/DigitalNatureGroup/CGH-autodiff) offers scripts for an automatic differentiation-based algorithm [@yamamoto2021gradient] for generating DMD hologram gerenation. [`SLMLayout`](https://github.com/wavefrontshaping/Layout) is a package focusing on wavefront shaping with macopixel method [@matthes2019optical] with DMD. 
-
[…]
