### 2024-06-26 a2cc947
better opening sentence in pkg comparison paragraph
