+++ paper.md
@@ -50,7 +50,7 @@ Secondly, holographic beam shaping can actively correct optical aberrations in t
 
 At the time of writing, the `pySLM2` package, as detailed in this manuscript, has been used in the trapped ion quantum information processing research [@shih2021reprogrammable;@motlakunta2023preserving;@kotibhaskar2023programmable]. The authors believe that the package will benefit a broader community of researchers and engineers by offering turnkey solutions for applying holographic beam shaping to their work. Moreover, the primitives included in the package can assist researchers in rapidly prototyping new hologram generation algorithms.
 
-There are existing open-source packages available that offer functionalities similar to `pySLM2`. [`SLMLayout`](https://github.com/wavefrontshaping/Layout) is a package focusing on wavefront shaping with macopixel method [@matthes2019optical] with DMD. [`CGH-diff`](https://github.com/DigitalNatureGroup/CGH-autodiff) offers scripts for an automatic differentiation-based algorithm [@yamamoto2021gradient] for generating DMD hologram gerenation. The most comparable package is [`slmsuite`](https://slmsuite.readthedocs.io/), which offers a comprehensive suite of hologram algorithms tailored for phase-only Spatial Light Modulators (SLMs), supported by GPU acceleration via CuPy [@nishino2017cupy], and includes hardware control capabilities for Liquid Crystal on Silicon (LCoS) SLMs. In contrast, `pySLM2` encompasses hardware controls for multiple popular commercial DMD and SLM models, implementations of four different algorithms for Fourier hologram generation, and simulation tools for visualizing the generated hologram, supported by GPU acceleration via `TensorFlow`. The authors believe that the addition of `pySLM2` to the community can facilitate researchers in applying desired Fourier holographic wavefront shaping ideas on actual machines with fewer hardware hurdles.
+There are existing open-source packages available that specilizes in different levels of holographic beam shaping with SLMs. [`SLMLayout`](https://github.com/wavefrontshaping/Layout) is a package focusing on wavefront shaping with macopixel method [@matthes2019optical] with DMD. [`CGH-diff`](https://github.com/DigitalNatureGroup/CGH-autodiff) offers scripts for an automatic differentiation-based algorithm [@yamamoto2021gradient] for generating DMD hologram gerenation. The most comparable package is [`slmsuite`](https://slmsuite.readthedocs.io/), which offers a comprehensive suite of hologram algorithms tailored for phase-only Spatial Light Modulators (SLMs), supported by GPU acceleration via CuPy [@nishino2017cupy], and includes hardware control capabilities for Liquid Crystal on Silicon (LCoS) SLMs. In contrast, `pySLM2` encompasses hardware controls for multiple popular commercial DMD and SLM models, implementations of four different algorithms for Fourier hologram generation, and simulation tools for visualizing the generated hologram, supported by GPU acceleration via `TensorFlow`. The authors believe that the addition of `pySLM2` to the community can facilitate researchers in applying desired Fourier holographic wavefront shaping ideas on actual machines with fewer hardware hurdles.
 
 # Fourier Holography Basics
 
