### 2024-07-05 #6315 @HaoZeke
Thanks @ldes89150, happy fourth ^_^

@brandondube, @sidihamady and @wavefrontshaping, could you please take a look and confirm that the issues raised were addressed?
