### 2024-07-05 #6315 @wavefrontshaping
All issues have been addressed, on my side all is good except for some typos to correct (here are some I found, pretty sure I missed some).

line 60: `macopixel` -> `macropixel`
line 60: `DMD` -> `DMDs`
line 61-62: `for generating DMD hologram gerenation` -> `for DMD hologram gerenation`
line 89: `two type of` -> `two types of`
line 89: I thing all the nouns should be plurals (DMDs, etc.)
line 108: `As illustrated in Fig. Figure 1`  -> `As illustrated in Figure 1`
