### 2024-07-05 #6315 @sidihamady
The authors addessed the issues, modulo the typos noted by my colleague @wavefrontshaping
