### 2024-07-07 QITI/pySLM2#16 @brandondube
I have read the current draft of the paper.  Overall, it is much improved - thank you!  A few areas I would request improvement:

> ... It has been shown that residual aberrations can be corrected to less than 𝜆/20 root-mean-square (RMS) which is smaller than most of the manufacturing tolerances of optical elements.

This is not really true.  For example, a "standard" off-the-shelf lens is finished to an irregularity of a quarter wave PV (quoted) and more like lambda/5 to 6 PV if you actually do incoming inspection of the components.  With the rule of thumb that RMS ~ 1/4 to 1/5 of PV, lambda/20 RMS is about as good as a "standard" off the shelf component, and worse than a "precision" or "high precision" component.  I think it would serve your purpose better to say `...lambda/20, which satisfies the Maréchal criteria`, which nobody can argue with

Your paper now features a brief survey of other SLM related packages on github, which is good.  However, I do not think it does an adequate job of relating the package to other works; it reads as though you did a search on Github, and listed approximately the first sentence about each package you found.  For example, CGH-diff also uses tensorflow, and so in the same way that your library supports CPU or GPU, so does that one.  You highlight a few features your library has,

1. Support for multiple (sic: two) vendors
2. Multiple algorithms
3. The user can view the far-field pattern of the hologram with your library
4. You have GPU acceleration

Regarding (1), I find it strained to say that two vendors "contrasts" to one when comparing to the other works in the field.  r.e (3), this is possible with all of the other libraries you referenced, and item (4) is a feature of 2/3 of the libraries you referenced.  These seem to be largely non-distinguishing features.

Your library is not required to be completely unique, filling a niche no other library does, but I think you should do a better job noting that most of the other libraries in the field are rather comparable.
