### 2024-07-07 QITI/pySLM2#13 @brandondube
While I still think your depiction of the achieved GPU acceleration with tensorflow is overly optimistic and a significantly greater acceleration is possible with a bit more effort, I am satisfied that you have made a significant effort to improve the clarity of communication in this area.
