### 2024-07-08 1c8c345
fixed typos in paper
