### 2024-07-08 e5ca996
fixed typos in pySLM2
