+++ docs/algorithm.md
@@ -17,7 +17,7 @@ For DMDs, use the method keyword argument in the `slm.calculate_dmd_state` funct
 ### `pySLM2.slm.LCOS_SLM.calculate_hologram`
 
 - `gs`: The Gerchberg-Saxton algorithm is an iterative phase retrieval process. The refined phase profile ideally would converge to the target phase profile after enough iterations. However, the convergence can be slow and is not guaranteed. [[1]](#ref1)  When using this method, the user can use kwarg `N` to define number of iterations desired. 
-- `mraf`: The Mixed-Region Amplitude Freedom algorithm is an improved adpataion of the `gs` algorihtms which iteratively refines the phase profile within a specified signal window region. It improves the convergence within the signal window at the sacrifice of phase control in the outside region. [[2]](#ref2)[[7]](#ref7) When using this method, in addtion to `N`, user can also define `signal_window` as the region that needs correction, as well as `mixing_factor`, which is to adjust how much the outside region weighs in the correction. When `mixing_factor = 1`, no outside region is being weighted and correctino is only applied for signal window region. Suggested in [[2]](#ref2), the `mixing_factor` set to `0.4` typically yields the lowest RMS error.
+- `mraf`: The Mixed-Region Amplitude Freedom algorithm is an improved adaption of the `gs` algorithms which iteratively refines the phase profile within a specified signal window region. It improves the convergence within the signal window at the sacrifice of phase control in the outside region. [[2]](#ref2)[[7]](#ref7) When using this method, in addition to `N`, user can also define `signal_window` as the region that needs correction, as well as `mixing_factor`, which is to adjust how much the outside region weighs in the correction. When `mixing_factor = 1`, no outside region is being weighted and correction is only applied for signal window region. Suggested in [[2]](#ref2), the `mixing_factor` set to `0.4` typically yields the lowest RMS error.
 
 ### `pySLM2.slm.DMD.calculate_dmd_state`
 
