+++ paper.md
@@ -46,11 +46,17 @@ High-quality optical controls are crucial for numerous scientific and engineerin
 Holographic beam shaping using SLMs provides a way for precise and adaptive optical controls. Compared to using conventional optical elements, holographic beam shaping has several advantages. Firstly, it can generate arbitrary beam profiles that are challenging to create with standard optical elements. For example, the Laguerre-Gaussian beam with a non-zero azimuthal index (often referred to as a doughnut beam), which can be used to trap atoms in a tube-like potential [@kuga1997novel], apply angular momentum to Bose-Einstein Condensate [@andersen2006quantized], or achieve super-resolution imaging [@qian2021super;@drechsler2021optical].
 
 
-Secondly, holographic beam shaping can actively correct optical aberrations in the system, thereby achieving diffraction-limited performance. This enables the faithful production of target beam profiles with high accuracy. It has been shown that residual aberrations can be corrected to less than $\lambda/20$ root-mean-square (RMS) [@shih2021reprogrammable;@zupancic2016ultra], which is smaller than most of the manufacturing tolerances of optical elements.
+Secondly, holographic beam shaping can actively correct cumulative optical aberrations in the system arising from almost inevitable causes, such as surface irregularities, misalignment, and imperfect lens curvature, thereby achieving diffraction-limited performance at the target image plane. This enables the faithful production of target beam profiles with high accuracy, relieving the stringent requirement on optics quality and alignment precision. It has been shown that residual wavefront aberrations can be corrected to less than $\lambda/20$ root-mean-square (RMS) [@shih2021reprogrammable;@zupancic2016ultra], which meets the Maréchal criteria (wavefront RMS error < $\lambda/14$) for optical performance. 
+<!-- 
+It is important to note that our method's strength lies in its ability to achieve this level of correction regardless of the initial beam quality. Even when cumulative aberrations from various optical components exceed λ/20, our approach ensures that the final corrected beam meets or exceeds this threshold, providing high-quality performance that is comparable to or better than what can be achieved with high-precision optical elements. -->
+
+
 
 At the time of writing, the `pySLM2` package, as detailed in this manuscript, has been used in the trapped ion quantum information processing research [@shih2021reprogrammable;@motlakunta2023preserving;@kotibhaskar2023programmable]. The authors believe that the package will benefit a broader community of researchers and engineers by offering turnkey solutions for applying holographic beam shaping to their work. Moreover, the primitives included in the package can assist researchers in rapidly prototyping new hologram generation algorithms.
 
[…]
