### 2024-08-04 8bbe43e
update the reference and discussion on packages comparision
