### 2024-08-04 b19bd6e
Addressing package similarities and aberration correction scale issues (#29)
* modified package review paragraph, modified aberration correction paragraph

* update the reference and discussion on packages comparision

---------

Co-authored-by: gilbert <ldes89150@gmail.com>
