+++ paper.bib
@@ -39,12 +39,6 @@
   year={1972}
 }
 
-@article{motlakunta2023preserving,
-  title={Preserving a qubit during adjacent measurements at a few micrometers distance},
-  author={Motlakunta, Sainath and Kotibhaskar, Nikhil and Shih, Chung-You and Vogliano, Anthony and Mclaren, Darian and Hahn, Lewis and Zhu, Jingwen and Habl{\"u}tzel, Roland and Islam, Rajibul},
-  journal={arXiv preprint arXiv:2306.03075},
-  year={2023}
-}
 
 @software{sebastien_m_popoff_2022_6121191,
   author       = {Sébastien M. Popoff and
@@ -212,4 +206,27 @@
   pages={2176--2190},
   year={2008},
   publisher={Optica Publishing Group}
+}
+
+@article{motlakunta2024,
+author={Motlakunta, Sainath
+and Kotibhaskar, Nikhil
+and Shih, Chung-You
+and Vogliano, Anthony
+and McLaren, Darian
+and Hahn, Lewis
+and Zhu, Jingwen
+and Habl{\"u}tzel, Roland
+and Islam, Rajibul},
+title={Preserving a qubit during state-destroying operations on an adjacent qubit at a few micrometers distance},
+journal={Nature Communications},
+year={2024},
+month={Aug},
+day={03},
+volume={15},
+number={1},
+pages={6575},
+issn={2041-1723},
+doi={10.1038/s41467-024-50864-2},
+url={https://doi.org/10.1038/s41467-024-50864-2}
 }
