### 2024-08-04 897054c
add acknowledgments for funding supports
