### 2024-08-04 210480e
plural forms typos corrected
