+++ paper.md
@@ -51,9 +51,9 @@ Secondly, holographic beam shaping can actively correct cumulative optical aberr
 
 At the time of writing, the `pySLM2` package, as detailed in this manuscript, has been used in the trapped ion quantum information processing researches [@shih2021reprogrammable;@motlakunta2024;@kotibhaskar2023programmable]. The authors believe that the package will benefit a broader community of researchers and engineers by offering turnkey solutions for applying holographic beam shaping to their work. Moreover, the primitives included in the package can assist researchers in rapidly prototyping new hologram generation algorithms.
 
-There are existing open-source packages available that specialize in different levels of holographic beam shaping with SLMs. For example, [`SLMLayout`](https://github.com/wavefrontshaping/Layout) is a package focusing on wavefront shaping with macropixel method [@matthes2019optical] with DMDs. [`CGH-diff`](https://github.com/DigitalNatureGroup/CGH-autodiff) offers scripts for an automatic differentiation-based algorithm [@yamamoto2021gradient] for DMD hologram generation. [`slmsuite`](https://slmsuite.readthedocs.io/) offers a comprehensive suite of hologram algorithms tailored for phase-only Spatial Light Modulators (SLMs), supported by GPU acceleration via CuPy [@nishino2017cupy], and includes hardware control capabilities for Liquid Crystal on Silicon (LCoS) SLMs. 
+There are existing open-source packages available that specialize in different levels of holographic beam shaping with SLMs. For example, [`SLMLayout`](https://github.com/wavefrontshaping/Layout) is a package focusing on wavefront shaping with macropixel method [@matthes2019optical] with DMDs. [`CGH-diff`](https://github.com/DigitalNatureGroup/CGH-autodiff) offers `Tensorflow` scripts for an automatic differentiation-based algorithm [@yamamoto2021gradient] for DMD hologram generation. [`slmsuite`](https://slmsuite.readthedocs.io/) offers a comprehensive suite of hologram algorithms tailored for phase-only Spatial Light Modulators (SLMs), supported by GPU acceleration via `CuPy` [@nishino2017cupy], and includes hardware control capabilities for Liquid Crystal on Silicon (LCoS) SLMs. 
 
-`pySLM2` implements hologram generation algorithms and primitives leveraging TensorFlow for GPU acceleration. Some algorithms, including the algorithms used in authors' prior works[@shih2021reprogrammable;@motlakunta2024] in the field of quantum information processing, are only available in `pySLM2` at the time of writing. In terms of hardware controls, `pySLM2` offers a unified API for programming DMDs from different vendors. 
+`pySLM2` implements hologram generation algorithms and primitives leveraging `TensorFlow` for GPU acceleration. Some algorithms, including the algorithms used in authors' prior works[@shih2021reprogrammable;@motlakunta2024] in the field of quantum information processing, are only available in `pySLM2` at the time of writing. In terms of hardware controls, `pySLM2` offers a unified API for programming DMDs from different vendors. 
 
 We hope that the addition of `pySLM2` to the community will facilitate researchers in applying Fourier holographic beam shaping ideas on various machines with fewer hardware hurdles.
 
