### 2024-10-19 #6315 @ldes89150
Hey @

> @HaoZeke @brandondube @wavefrontshaping @sidihamady
>
> My apologies for the delayed response.
>
> It took us some time to push a work we cited in the paper through the finishing line. https://www.nature.com/articles/s41467-024-50864-2
>
> Now, the paper has been updated to address the new issues, and it's reviewed by all the authors.
>
> Thank you all for your valuable time in helping with the review!

Hey @brandondube ! I've already updated the paper here to address the issues here.
My apologies if there is any miscommunication.
