### 2024-10-19 #6315 @sidihamady
> @wavefrontshaping and @sidihamady could you provide an update on the review progress as well?

Hello, I already indicated that the authors addessed the issues, in my comment posted July 5.
