### 2024-10-19 #6315 @HaoZeke
Ah my apologies, @sidihamady could you finish marking the checklist [items here](https://github.com/openjournals/joss-reviews/issues/6315#issuecomment-1935646506)?

@wavefrontshaping thanks for the quick turn around! I see somehow the checklist wasn't generated for you at all (my bad). Since the review is complete, could you generate and check off the items by first running:

and then ticking things off?

@brandondube, please let us know if the clarifications are insufficient.
