### 2024-10-24 QITI/pySLM2#16 @brandondube
I am satisfied with the changes to the manuscript
