### 2024-10-24 f3721f8
adding urls for references except for one article
