+++ paper.bib
@@ -6,7 +6,8 @@
   number={1},
   pages={57},
   year={2021},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  url={https://doi.org/10.1038/s41534-021-00396-0}
 }
 
 @article{zupancic2016ultra,
@@ -17,7 +18,8 @@
   number={13},
   pages={13881--13893},
   year={2016},
-  publisher={Optica Publishing Group}
+  publisher={Optica Publishing Group},
+  url={https://doi.org/10.1364/OE.24.013881}
 }
 
 @incollection{lee1978iii,
@@ -27,7 +29,8 @@
   volume={16},
   pages={119--232},
   year={1978},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  url={https://doi.org/10.1016/S0079-6638(08)70072-6}
 }
 
 @article{gerhberg1972practical,
@@ -36,7 +39,8 @@
   journal={Optik},
   volume={35},
   pages={237--246},
-  year={1972}
+  year={1972},
+  url={https://api.semanticscholar.org/CorpusID:55691159}
 }
 
 
@@ -62,7 +66,8 @@
   number={1},
   pages={721},
   year={2012},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  url = {https://doi.org/10.1038/srep00721},
 }
 
 @article{qian2021super,
@@ -73,7 +78,8 @@
   number={26},
   pages={263603},
   year={2021},
-  publisher={APS}
+  publisher={APS},
+  url={https://doi.org/10.1103/PhysRevLett.127.263603}
 }
 
 @article{drechsler2021optical,
@@ -84,14 +90,16 @@
   number={14},
   pages={143602},
   year={2021},
-  publisher={APS}
+  publisher={APS},
+  url={https://doi.org/10.1103/PhysRevLett.127.143602}
 }
 
 @article{kotibhaskar2023programmable,
   title={Programmable XY-type couplings through parallel spin-dependent forces on the same trapped ion motional modes},
   author={Kotibhaskar, Nikhil and Shih, Chung-You and Motlakunta, Sainath and Vogliano, Anthony and Hahn, Lewis and Chen, Yu-Ting and Islam, Rajibul},
   journal={arXiv preprint arXiv:2307.04922},
-  year={2023}
+  year={2023},
+  url={https://doi.org/10.1103/PhysRevResearch.6.033038}
 }
 
 @article{andersen2006quantized,
@@ -102,7 +110,8 @@
   number={17},
   pages={170406},
   year={2006},
-  publisher={APS}
+  publisher={APS},
+  url={https://doi.org/10.1103/PhysRevLett.97.170406}
 }
 
 @article{kuga1997novel,
@@ -113,7 +122,9 @@
   number={25},
   pages={4713},
   year={1997},
-  publisher={APS}
+  publisher={APS},
+  url={https://doi.org/10.1103/PhysRevLett.78.4713}
+
 }
 
 @article{obata2010multi,
@@ -124,7 +135,8 @@
   number={16},
   pages={17193--17200},
   year={2010},
-  publisher={Optica Publishing Group}
+  publisher={Optica Publishing Group},
+  url={https://doi.org/10.1364/OE.18.017193}
 }
 
 @article{kim2019large,
@@ -135,7 +147,9 @@
   number={12},
   pages={3178--3181},
   year={2019},
-  publisher={Optica Publishing Group}
+  publisher={Optica Publishing Group},
+  url={https://doi.org/10.1364/OL.44.003178}
+
 }
 
 @article{islam2015measuring,
@@ -146,7 +160,8 @@
   number={7580},
   pages={77--83},
   year={2015},
[…]
