+++ paper.bib
@@ -7,6 +7,7 @@
   pages={57},
   year={2021},
   publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41534-021-00396-0},
   url={https://doi.org/10.1038/s41534-021-00396-0}
 }
 
@@ -19,6 +20,7 @@
   pages={13881--13893},
   year={2016},
   publisher={Optica Publishing Group},
+  doi={10.1364/OE.24.013881},
   url={https://doi.org/10.1364/OE.24.013881}
 }
 
@@ -30,6 +32,7 @@
   pages={119--232},
   year={1978},
   publisher={Elsevier},
+  doi={10.1016/S0079-6638(08)70072-6},
   url={https://doi.org/10.1016/S0079-6638(08)70072-6}
 }
 
@@ -67,6 +70,7 @@
   pages={721},
   year={2012},
   publisher={Nature Publishing Group UK London},
+  doi={10.1038/srep00721},
   url = {https://doi.org/10.1038/srep00721},
 }
 
@@ -79,6 +83,7 @@
   pages={263603},
   year={2021},
   publisher={APS},
+  doi={10.1103/PhysRevLett.127.263603},
   url={https://doi.org/10.1103/PhysRevLett.127.263603}
 }
 
@@ -91,6 +96,7 @@
   pages={143602},
   year={2021},
   publisher={APS},
+  doi={10.1103/PhysRevLett.127.143602},
   url={https://doi.org/10.1103/PhysRevLett.127.143602}
 }
 
@@ -99,6 +105,7 @@
   author={Kotibhaskar, Nikhil and Shih, Chung-You and Motlakunta, Sainath and Vogliano, Anthony and Hahn, Lewis and Chen, Yu-Ting and Islam, Rajibul},
   journal={arXiv preprint arXiv:2307.04922},
   year={2023},
+  doi={10.1103/PhysRevResearch.6.033038},
   url={https://doi.org/10.1103/PhysRevResearch.6.033038}
 }
 
@@ -111,6 +118,7 @@
   pages={170406},
   year={2006},
   publisher={APS},
+  doi={10.1103/PhysRevLett.97.170406},
   url={https://doi.org/10.1103/PhysRevLett.97.170406}
 }
 
@@ -123,6 +131,7 @@
   pages={4713},
   year={1997},
   publisher={APS},
+  doi={10.1103/PhysRevLett.78.4713},
   url={https://doi.org/10.1103/PhysRevLett.78.4713}
 
 }
@@ -136,6 +145,7 @@
   pages={17193--17200},
   year={2010},
   publisher={Optica Publishing Group},
+  doi={10.1364/OE.18.017193},
   url={https://doi.org/10.1364/OE.18.017193}
 }
 
@@ -148,6 +158,7 @@
   pages={3178--3181},
   year={2019},
   publisher={Optica Publishing Group},
+  doi= {10.1364/OL.44.003178},
   url={https://doi.org/10.1364/OL.44.003178}
 
 }
@@ -161,6 +172,7 @@
   pages={77--83},
   year={2015},
   publisher={Nature Publishing Group UK London},
+  doi={10.1038/nature15750},
   url={https://doi.org/10.1038/nature15750}
 }
 
@@ -173,6 +185,7 @@
   pages={821--827},
   year={2008},
   publisher={Nature Publishing Group US New York},
+  doi={10.1038/nmeth.1241},
   url={https://doi.org/10.1038/nmeth.1241}
 }
 
@@ -201,6 +214,7 @@
   pages={DTh7C--1},
   year={2021},
   organization={Optica Publishing Group},
+  doi={10.1364/DH.2021.DTh7C.1},
[…]
