### 2024-10-24 #6315 @HaoZeke
@marvelousmonicaaa not quite there, but almost; for CuPy you'll want the `bib` from the repo:

```
@inproceedings{cupy_learningsys2017,
  author       = "Okuta, Ryosuke and Unno, Yuya and Nishino, Daisuke and Hido, Shohei and Loomis, Crissman",
  title        = "CuPy: A NumPy-Compatible Library for NVIDIA GPU Calculations",
  booktitle    = "Proceedings of Workshop on Machine Learning Systems (LearningSys) in The Thirty-first Annual Conference on Neural Information Processing Systems (NIPS)",
  year         = "2017",
  url          = "http://learningsys.org/nips17/assets/papers/paper_16.pdf"
}
```

For the Hartmann screen, maybe the DOI should just be of the program? https://doi.org/10.1364/JOSA.61.000648

Do you have a PDF for that? I couldn't find one. The Optik one is fine, they only digitized from 2001 so there's no DOI at all..
