### 2024-10-24 840d422
update cupy bib and hartmann ref
