+++ paper.bib
@@ -194,17 +194,17 @@
   author={Shack, Roland V},
   booktitle={Spring Meeting of Optical Society of America, 1971},
   volume={656},
-  year={1971}
+  year={1971},
+  doi={10.1364/JOSA.61.000648},
+  url={https://doi.org/10.1364/JOSA.61.000648}
 }
 
-@article{nishino2017cupy,
-  title={Cupy: A numpy-compatible library for nvidia gpu calculations},
-  author={Nishino, ROYUD and Loomis, Shohei Hido Crissman},
-  journal={31st confernce on neural information processing systems},
-  volume={151},
-  number={7},
-  year={2017},
-  url={http://learningsys.org/nips17/assets/papers/paper_16.pdf}
+@inproceedings{cupy_learningsys2017,
+  author       = "Okuta, Ryosuke and Unno, Yuya and Nishino, Daisuke and Hido, Shohei and Loomis, Crissman",
+  title        = "CuPy: A NumPy-Compatible Library for NVIDIA GPU Calculations",
+  booktitle    = "Proceedings of Workshop on Machine Learning Systems (LearningSys) in The Thirty-first Annual Conference on Neural Information Processing Systems (NIPS)",
+  year         = "2017",
+  url          = "http://learningsys.org/nips17/assets/papers/paper_16.pdf"
 }
 
 @inproceedings{yamamoto2021gradient,
