+++ paper.bib
@@ -43,7 +43,7 @@
   volume={35},
   pages={237--246},
   year={1972},
-  url={https://api.semanticscholar.org/CorpusID:55691159}
+  url={https://web.archive.org/web/20220505015731/http://www.u.arizona.edu/~ppoon/GerchbergandSaxton1972.pdf}
 }
 
 
