### 2024-10-25 26b2a5d
add another hartmann sensor reference and updated in paper.md
