+++ paper.bib
@@ -199,6 +199,21 @@
   url={https://doi.org/10.1364/JOSA.61.000648}
 }
 
+@misc{Paschottashack_hartmann_wavefront_sensors},
+  author = {Paschotta, R.},
+  publisher = {RP Photonics AG},
+  title = {Shack–Hartmann Wavefront Sensors},
+  abstract = {Shack--Hartmann sensors measure wavefront shapes and optical amplitude distributions. They are used in adaptive optics and laser beam characterization.},
+  year = {},
+  month = {},
+  howpublished = {RP Photonics Encyclopedia},
+  note = {{Available online at \url{https://www.rp-photonics.com/shack_hartmann_wavefront_sensors.html}},
+  url = {https://www.rp-photonics.com/shack_hartmann_wavefront_sensors.html},
+  doi = {10.61835/jcv" title="This link will reload the current page.},
+  urldate = {2024-10-25},
+  keywords = {Shack–Hartmann wavefront sensors}
+}
+
 @inproceedings{cupy_learningsys2017,
   author       = "Okuta, Ryosuke and Unno, Yuya and Nishino, Daisuke and Hido, Shohei and Loomis, Crissman",
   title        = "CuPy: A NumPy-Compatible Library for NVIDIA GPU Calculations",
