+++ paper.bib
@@ -199,7 +199,7 @@
   url={https://doi.org/10.1364/JOSA.61.000648}
 }
 
-@misc{Paschottashack_hartmann_wavefront_sensors},
+@misc{Paschottashack_hartmann_wavefront_sensors,
   author = {Paschotta, R.},
   publisher = {RP Photonics AG},
   title = {Shack–Hartmann Wavefront Sensors},
@@ -207,7 +207,7 @@
   year = {},
   month = {},
   howpublished = {RP Photonics Encyclopedia},
-  note = {{Available online at \url{https://www.rp-photonics.com/shack_hartmann_wavefront_sensors.html}},
+  note = {{Available online at \url{https://www.rp-photonics.com/shack_hartmann_wavefront_sensors.html}}},
   url = {https://www.rp-photonics.com/shack_hartmann_wavefront_sensors.html},
   doi = {10.61835/jcv" title="This link will reload the current page.},
   urldate = {2024-10-25},
