### 2024-10-25 #6315 @marvelousmonicaaa
> > > @marvelousmonicaaa not quite there, but almost; for CuPy you'll want the `bib` from the repo:
> > > ```
> > > @inproceedings{cupy_learningsys2017,
> > >   author       = "Okuta, Ryosuke and Unno, Yuya and Nishino, Daisuke and Hido, Shohei and Loomis, Crissman",
> > >   title        = "CuPy: A NumPy-Compatible Library for NVIDIA GPU Calculations",
> > >   booktitle    = "Proceedings of Workshop on Machine Learning Systems (LearningSys) in The Thirty-first Annual Conference on Neural Information Processing Systems (NIPS)",
> > >   year         = "2017",
> > >   url          = "http://learningsys.org/nips17/assets/papers/paper_16.pdf"
> > > }
> > > ```
> > >
> > >
> > >
> > >
> > >
> > >
> > >
> > >
> > >
> > >
> > >
> > > For the Hartmann screen, maybe the DOI should just be of the program? https://doi.org/10.1364/JOSA.61.000648
> > > Do you have a PDF for that and the Optik paper? I couldn't find one.
> > > The Optik one is fine without a DOI, Elsevier only [digitized from 2001](https://www.sciencedirect.com/journal/optik/issues?page=2) so there's no DOI at all..
> >
> >
> > @HaoZeke thanks for your suggestions. I updated the cupy bib from its repo and updated the link to JOSA for that Hartmann screen one. as for the optik one, i can't find digital copies anywhere for that article.
>
> Makes sense, but also; have you read the listed papers (e.g. perhaps in a library archive)? If not, perhaps it might be best to defer to a citation which can be checked for accuracy (and for posterity)?

Hi @HaoZeke, we actually found the scanned pdf version for this one. Now it is updated in the bib
