### 2024-10-25 ea898a4
fix the new link's doi line
