+++ paper.bib
@@ -209,7 +209,7 @@
   howpublished = {RP Photonics Encyclopedia},
   note = {{Available online at \url{https://www.rp-photonics.com/shack_hartmann_wavefront_sensors.html}}},
   url = {https://www.rp-photonics.com/shack_hartmann_wavefront_sensors.html},
-  doi = {10.61835/jcv" title="This link will reload the current page.},
+  doi = {10.61835/jcv},
   urldate = {2024-10-25},
   keywords = {Shack–Hartmann wavefront sensors}
 }
