### 2024-10-29 #6315 @HaoZeke
@marvelousmonicaaa could you make a new release and put it on zenodo as mentioned here? https://github.com/openjournals/joss-reviews/issues/6315#issuecomment-2434069127
