### 2024-11-01 bfa0266
add doi in README
