### 2024-11-05 #6315 @HaoZeke
@marvelousmonicaaa could you update the reference for cupy_learningsys2017 ?
