### 2024-11-09 350ce29
fixed reference label in paper.md
