### 2024-11-09 #6315 @marvelousmonicaaa
> @marvelousmonicaaa could you update the reference by merging [QITI/pySLM2#30](https://github.com/QITI/pySLM2/pull/30)?

Hey @HaoZeke, I've merge it. Thanks! Hopefully there's no other issues with the reference!

May I ask what tools you use to check references as editor? Maybe next time we can do the checks ourselves.
