### 2024-11-09 #6315 @HaoZeke
@marvelousmonicaaa generally I was using `editorialbot` commands, although that particular issue (wrong labels) should show up if you tried to build the `.tex` file locally.
