### 2024-11-15 #6315 @kyleniemeyer
@ldes89150 @marvelousmonicaaa regarding the paper, a few things to fix:
- figure mentions in the paper currently appear as "Fig. Figure 1" and "Fig. Figure 2". Please remove the unnecessary "Fig." occurences in the text.
- The paper mentions depending on TensorFlow, but I don't see a citation for that. Please add a citation for that and any other dependencies.
- When you have instances of author names in the text for a citation, please use the `@author:2001` citation command format (without square brackets) to get the author name in the text like "Author et al. (2001)". (Such as in the sentence "For a detailed description of the latter method, one can refer to Shih et al. (Shih et al., 2021)." -> "For a detailed description of the latter method, one can refer to Shih et al. (2021).")
