### 2024-11-26 e1310f4
fix: duplicate fig.
