### 2024-11-26 e05c17b
doc: add tensorflow citation
