+++ paper.bib
@@ -281,3 +281,14 @@ issn={2041-1723},
 doi={10.1038/s41467-024-50864-2},
 url={https://doi.org/10.1038/s41467-024-50864-2}
 }
+
+@software{tensorflow_developers_2024_13989084,
+  author       = {TensorFlow Developers},
+  title        = {TensorFlow},
+  month        = oct,
+  year         = 2024,
+  publisher    = {Zenodo},
+  version      = {v2.18.0},
+  doi          = {10.5281/zenodo.13989084},
+  url          = {https://doi.org/10.5281/zenodo.13989084}
+}
