### 2024-11-26 e137896
fix:  use the @author:2001 citation command
