### 2024-11-26 31e29b2
fix: author citation command
