### 2024-11-26 3a73b59
doc: add tensorflow hyperlink
