### 2024-12-03 #6315 @marvelousmonicaaa
Hi @kyleniemeyer, sorry about the delay. We have fixed most of the issues now. Just one thing to confirm, I noticed the DOI number in the pdf generated by editoialbot is a '10.xxxxx/draft". I wonder if that is autogenerated DOI for this paper and it will be an actual number after it is published?
