### 2024-12-05 #6315 @kyleniemeyer
@marvelousmonicaaa that paper DOI is a placeholder, it will be updated after accepting.

I only see one minor remaining issue: the author field is incorrect for the new Tensorflow citation (appearing as "Developers, T."). Anything that isn't a person's name needs to be protected with extra `{ }`, otherwise it gets interpreted (and abbreviated) like a name.
