### 2024-12-09 cdd4130
add note to tensorflow ref
